% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argMatch.R
\name{argMatch}
\alias{argMatch}
\title{Updates the Shiny app options if the function arguments change}
\usage{
argMatch(fxnName, justNames = FALSE)
}
\arguments{
\item{fxnName}{String. The name of the function from which arguments should
be drawn.}

\item{justNames}{Logical. Default is FALSE. Should only argument names be returned}
}
\value{
List of numericInput expressions.
}
\description{
Updates the Shiny app options if the function arguments change
}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu}
}
