% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPower.R
\docType{package}
\name{clusterPower}
\alias{clusterPower}
\title{clusterPower: doing various power analysis calculations for cluster randomized trials.}
\description{
clusterPower: doing various power analysis calculations for cluster randomized trials.
}
