% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.calc.R
\name{confint.calc}
\alias{confint.calc}
\title{Power simulations for cluster-randomized trials: get exact confidence intervals.}
\usage{
\method{confint}{calc}(nsim = nsim, alpha = alpha, p.val = p.val, names.power = names.power)
}
\arguments{
\item{nsim}{A scalar; the number of simulations chosen by the user.}

\item{alpha}{A numeric; the user-selected alpha cutoff.}

\item{p.val}{A vector of p-values.}

\item{names.power}{A vector of labels containing the names of each arm.}
}
\value{
A dataframe
\describe{
  \item{power.parms}{Exact confidence intervals produced using \code{binom.test()}}
}
}
\description{
This function can be called by some of the clusterPower functions to 
calculate the confidence intervals using binomial exact test.
}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu} and Ken Kleinman (\email{ken.kleinman@gmail.com})
}
