% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confintCalc.R
\name{confintCalc}
\alias{confintCalc}
\title{Power simulations for cluster-randomized trials: get exact confidence intervals.}
\usage{
confintCalc(alpha = alpha, p.val = p.val, nsim = nsim, multi = FALSE)
}
\arguments{
\item{alpha}{A numeric; the user-selected alpha cutoff.}

\item{p.val}{A vector of p-values.}
}
\value{
A dataframe
\describe{
  \item{power.parms}{Exact confidence intervals produced using \code{binom.test()}}
}
}
\description{
This function uses binom.test() to get exact
confidence intervals for the power in Monte Carlo power estimation.
It is called by some of the clusterPower functions.
}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu} and Ken Kleinman (\email{ken.kleinman@gmail.com})
}
