% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpa.count.R
\name{cpa.count}
\alias{cpa.count}
\title{Power calculations for simple cluster randomized trials, count outcome}
\usage{
cpa.count(
  alpha = 0.05,
  power = 0.8,
  n = NA,
  m = NA,
  r1 = NA,
  r2 = NA,
  CV = NA,
  r1inc = TRUE,
  tol = .Machine$double.eps^0.25
)
}
\arguments{
\item{alpha}{The level of significance of the test, the probability of a
Type I error.}

\item{power}{The power of the test, 1 minus the probability of a Type II
error.}

\item{r1}{The expected mean event rate per unit time in the treatment group.}

\item{r2}{The mean event rate per unit time in the control group.}

\item{r1inc}{Logical indicating if r1 is expected to be greater than r2. This is
only important to specify if one of r1 or r2 is NA.}

\item{tol}{Numerical tolerance used in root finding. The default provides
at least four significant digits.}

\item{nclusters}{The number of clusters per condition. It must be greater than 1.}

\item{py}{The number of person-years of observation per cluster.}

\item{cvb}{The between-cluster coefficient of variation.}
}
\value{
The computed argument.
}
\description{
Wrapper function that uses crtpwr.2rate to compute the power of a 
simple cluster randomized trial with a count outcome,
or determine parameters to obtain a target power. For additional
details, see ?crtpwr.2rate help page.
}
\section{Authors}{

Jonathan Moyer (\email{jon.moyer@gmail.com}), Ken Kleinman (\email{ken.kleinman@gmail.com})
}

\examples{
# Find the number of clusters per condition needed for a trial with alpha = 0.05, 
# power = 0.80, 10 person-years per cluster, rate in condition 1 of 0.10 
# and condition 2 of 0.20, and CVB = 0.10.
cpa.count(m=10, r1=0.10, r2=0.20, CV=0.10)
# 
# The result, showimg nclusters of greater than 24, suggests 25 clusters per
# condition should be used.

}
\references{
Donner A, Klar N. Design and Analysis of Cluster Randomization Trials in Health Research. Chichester, UK; 2009.

Hayes JR, Moulton LH. Cluster Randomized Trials. Boca Raton, FL: CRC Press; 2009.
}
