% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpa.irgtt.count.R
\name{cpa.irgtt.count}
\alias{cpa.irgtt.count}
\title{This help page is a stub. Equations do not yet exist for this type of analysis and outcome.
Try \code{cps.irgtt.count()} instead.}
\usage{
cpa.irgtt.count(...)
}
\arguments{
\item{...}{Any argument passed to the function.}
}
\value{
A helpful suggestion to use cps.irgtt.count() instead.
}
\description{
This help page is a stub. Equations do not yet exist for this type of analysis and outcome.
Try \code{cps.irgtt.count()} instead.
}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu}

Ken Kleinman (\email{ken.kleinman@gmail.com})
}
