% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpa.ma.binary.R
\name{cpa.ma.binary}
\alias{cpa.ma.binary}
\title{This help page is a stub. Equations do not yet exist for this type of analysis and outcome.
Try \code{cps.did.binary()} instead.}
\usage{
cpa.ma.binary(...)
}
\arguments{
\item{...}{Any argument passed to the function.}
}
\value{
A helpful suggestion to use cps.ma.binary() instead.
}
\description{
This help page is a stub. Equations do not yet exist for this type of analysis and outcome.
Try \code{cps.did.binary()} instead.
}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu}

Ken Kleinman (\email{ken.kleinman@gmail.com})
}
