% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpa.normal.R
\name{cpa.normal}
\alias{cpa.normal}
\title{Power calculations for simple cluster randomized trials, continuous outcome}
\usage{
cpa.normal(alpha = 0.05, power = 0.8, n = NA, m = NA, cv = 0,
  difference = NA, ICC = NA, vart = NA, method = c("taylor",
  "weighted"), tol = .Machine$double.eps^0.25)
}
\arguments{
\item{alpha}{The level of significance of the test, the probability of a
Type I error.}

\item{power}{The power of the test, 1 minus the probability of a Type II
error.}

\item{n}{The number of clusters per condition. It must be greater than 1.}

\item{m}{The mean of the cluster sizes, or a vector of cluster sizes for one arm.}

\item{cv}{The coefficient of variation of the cluster sizes. When \code{cv} = 0,
the clusters all have the same size.}

\item{difference}{The difference in condition means.}

\item{ICC}{The intraclass correlation.}

\item{vart}{The total variation of the outcome (the sum of within- and between-cluster variation).}

\item{method}{The method for calculating variance inflation due to unequal cluster
sizes. Either a method based on Taylor approximation of relative efficiency 
("taylor"), or weighting by cluster size ("weighted")}

\item{tol}{Numerical tolerance used in root finding. The default provides
at least four significant digits.}
}
\value{
The computed argument.
}
\description{
Wrapper function that uses crtpwr.2mean to compute the power of 
a simple cluster randomized trial with a continuous outcome,
or determine parameters to obtain a target power. See ?crtpwr.2mean 
for more details.
#' @section Authors:
Alexandria Sakrejda, Jonathan Moyer (\email{jon.moyer@gmail.com}), Ken Kleinman (\email{ken.kleinman@gmail.com})
}
\examples{
# Find the number of clusters per condition needed for a trial with alpha = .05, 
# power = 0.8, 10 observations per cluster, no variation in cluster size, a difference 
# of 1 unit,  icc = 0.1 and   a variance of five units.
cpa.normal(m=10, difference=1, ICC=.1, vart=5)
# 
# The result, showimg nclusters of greater than 15, suggests 16 clusters per 
# condition should be used.
}
\references{
Eldridge SM, Ukoumunne OC, Carlin JB. (2009) The Intra-Cluster Correlation
  Coefficient in Cluster Randomized Trials: A Review of Definitions. Int Stat Rev. 
  77: 378-394.

Eldridge SM, Ashby D, Kerry S. (2006) Sample size for cluster randomized
  trials: effect of coefficient of variation of cluster size and analysis method.
  Int J Epidemiol. 35(5):1292-300.

van Breukelen GJP, Candel MJJM, Berger MPF. (2007) Relative efficiency of
  unequal versus equal cluster sizes in cluster randomized and multicentre trials.
  Statist Med. 26:2589-2603.
}
