% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.did.binary.R
\name{cps.did.binary}
\alias{cps.did.binary}
\title{Power simulations for cluster-randomized trials: Difference in Difference, Binary Outcome.}
\usage{
cps.did.binary(nsim = NULL, nsubjects = NULL, nclusters = NULL,
  p.diff = NULL, p1 = NULL, p2 = NULL, or1 = NULL, or2 = NULL,
  or.diff = NULL, sigma_b0 = NULL, sigma_b1 = 0, alpha = 0.05,
  method = "glmm", quiet = TRUE, all.sim.data = FALSE)
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{nsubjects}{Number of subjects per cluster; accepts integer (required).}

\item{nclusters}{Number of clusters per treatment group; accepts integer (required).
At least 2 of the following 3 arguments must be specified when using expected probabilities:}

\item{p.diff}{Expected difference in probability of outcome between groups, defined as p.diff = p1 - p2
At least 2 of the following 3 arguments must be specified when using expected odds ratios:}

\item{p1}{Expected probability of outcome in non-treatment group}

\item{p2}{Expected probability of outcome in treatment group}

\item{or1}{Expected odds ratio for outcome in non-treatment group}

\item{or2}{Expected odds ratio for outcome in treatment group}

\item{or.diff}{Expected difference in odds ratio for outcome between groups, defined as or.diff = or1 - or2}

\item{sigma_b0}{Pre-treatment (time == 0) between-cluster variance; accepts numeric scalar (indicating equal 
between-cluster variances for both treatment groups) or a vector of length 2 specifying treatment-specific 
between-cluster variances}

\item{sigma_b1}{Post-treatment (time == 1) between-cluster variance; accepts numeric scalar (indicating equal 
between-cluster variances for both treatment groups) or a vector of length 2 specifying treatment-specific 
between-cluster variances. For data simulation, SIGMA_B1 is added to SIGMA_B0, such that if SIGMA_B0 = 5 
and SIGMA_B1 = 2, the between-cluster variance at time == 1 equals 7. Default = 0.}

\item{alpha}{Significance level. Default = 0.05}

\item{method}{Analytical method, either Generalized Linear Mixed Effects Model (GLMM) or Generalized Estimating Equation (GEE). Accepts c('glmm', 'gee') (required); default = 'glmm'.}

\item{quiet}{When set to FALSE, displays simulation start time and completion time. Default is TRUE.}

\item{all.sim.data}{Option to output list of all simulated datasets. Default = FALSE}
}
\description{
This function utilizes iterative simulations to determine 
approximate power for cluster-randomized controlled trials. Users 
can modify a variety of parameters to suit the simulations to their
desired experimental situation.
}
\details{
Runs the power simulation for difference in difference RCTs with binary outcomes.

Users must specify the desired number of simulations, number of subjects per 
cluster, number of clusters per treatment arm, pre-treatment between-cluster variance, 
and two of the following three terms: 
expected probability of outcome in non-treatment group, expected probability of 
outcome in treatment group, expected difference in probabilities between groups
; post-treatment between-cluster variance, significance level, analytic method, progress updates, 
and simulated data set output may also be specified.

The following equations are used to estimate intra-cluster correltation coefficients:
P_h: \deqn{ICC = \frac{\sigma_{b}}{\sigma_{b} + \pi^{2}/3}}
P_c: \deqn{ICC = \frac{P(Y_{ij} = 1, Y_{ih} = 1) - \pi_{j}\pi_{h}}{\sqrt{\pi_{j}(1 - \pi_{j})\pi_{h}(1 - \pi_{h})}}}
P_lmer: \deqn{ICC = \frac{\sigma_{b}}{\sigma_{b} + \sigma_{w}}}
}
\examples{
\dontrun{
did.binary.sim = cps.did.binary(nsim = 100, nsubjects = 50, nclusters = 6, 
                                p1 = 0.4, p2 = 0.2, sigma_b0 = 10, alpha = 0.05,
                                method = 'glmm', all.sim.data = FALSE)
}

}
\references{
Snjiders, T. & Bosker, R. Multilevel Analysis: an Introduction to Basic and Advanced Multilevel Modelling. London, 1999: Sage.

Elridge, S., Ukoumunne, O. & Carlin, J. The Intra-Cluster Correlation Coefficient in Cluster Randomized Trials: 
A Review of Definitions. International Statistical Review (2009), 77, 3, 378-394. doi: 10.1111/j.1751-5823.2009.00092.x
}
\author{
Alexander R. Bogdan
}
