% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.did.count.R
\name{cps.did.count}
\alias{cps.did.count}
\title{Power simulations for cluster-randomized trials: Difference in Difference, Count Outcome.}
\usage{
cps.did.count(nsim = NULL, nsubjects = NULL, nclusters = NULL,
  c1 = NULL, c2 = NULL, c.diff = NULL, sigma_b0 = NULL,
  sigma_b1 = 0, family = "poisson", analysis = "poisson",
  method = "glmm", alpha = 0.05, quiet = FALSE,
  all.sim.data = FALSE)
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{nsubjects}{Number of subjects per cluster; accepts integer (required).}

\item{nclusters}{Number of clusters per treatment group; accepts integer (required).
At least 2 of the following 3 arguments must be specified:}

\item{c1}{Expected outcome count in non-treatment group}

\item{c2}{Expected outcome count in treatment group}

\item{c.diff}{Expected difference in outcome count between groups, defined as c.diff = c1 - c2}

\item{sigma_b0}{Pre-treatment (time == 0) between-cluster variance; accepts numeric scalar (indicating equal 
between-cluster variances for both treatment groups) or a vector of length 2 specifying treatment-specific 
between-cluster variances}

\item{sigma_b1}{Post-treatment (time == 1) between-cluster variance; accepts numeric scalar (indicating equal 
between-cluster variances for both treatment groups) or a vector of length 2 specifying treatment-specific 
between-cluster variances. For data simulation, SIGMA_B1 is added to SIGMA_B0, such that if SIGMA_B0 = 5 
and SIGMA_B1 = 2, the between-cluster variance at time == 1 equals 7. Default = 0.}

\item{family}{Distribution from which responses are simulated. Accepts Poisson ('poisson') or negative binomial ('neg.binom') (required); default = 'poisson'}

\item{analysis}{Family used for regression; currently only applicable for GLMM. Accepts c('poisson', 'neg.binom') (required); default = 'poisson'}

\item{method}{Analytical method, either Generalized Linear Mixed Effects Model (GLMM) or Generalized Estimating Equation (GEE). Accepts c('glmm', 'gee') (required); default = 'glmm'}

\item{alpha}{Significance level for power estimation, accepts value between 0 - 1; default = 0.05}

\item{quiet}{When set to FALSE, displays simulation progress and estimated completion time. Default = FALSE.}

\item{all.sim.data}{Option to output list of all simulated datasets. Default = FALSE}
}
\value{
A list with the following components
  \item Character string indicating total number of simulations, 
  distribution of simulated data, and regression family
  \item Number of simulations
  \item Data frame with columns 'Power' (Estimated statistical power), 
               'lower.95.ci' (Lower 95% confidence interval bound), 
               'upper.95.ci' (Upper 95% confidence interval bound)
  \item Analytic method used for power estimation
  \item Data frame containing families for distribution and analysis of simulated data
  \item Significance level
  \item Vector containing user-defined cluster sizes
  \item Vector containing user-defined number of clusters
  \item Data frame reporting between-cluster variances at each time point for 
  each treatment group
  \item Vector containing expected counts and risk ratios based on user inputs
  \item Data frame with columns: 
                  'Period' (Pre/Post-treatment indicator), 
                  'Treatment' (Treatment group indicator), 
                  'Value' (Mean response value)
  \item Data frame with columns: 
                  'Estimate' (Estimate of treatment effect for a given simulation), 
                  'Std.Err' (Standard error for treatment effect estimate), 
                  'Test.statistic' (z-value (for GLMM) or Wald statistic (for GEE)), 
                  'p.value', 
                  'converge' (Did simulated model converge?), 
                  'sig.val' (Is p-value less than alpha?)
  \item List of data frames, each containing: 
                  'y' (Simulated response value), 
                  'trt' (Indicator for treatment group), 
                  'clust' (Indicator for cluster), 
                  'period' (Indicator for time point)
}
\description{
This function utilizes iterative simulations to determine 
approximate power for cluster-randomized controlled trials. Users 
can modify a variety of parameters to suit the simulations to their
desired experimental situation.
}
\details{
Runs power simulations for difference in difference cluster randomized control trials using count outcomes

Users must specify the desired number of simulations, number of subjects per 
cluster, number of clusters per treatment arm, between-cluster variance, 
two of the following: expected count in non-treatment group, expected count 
in treatment group, difference in counts between groups; significance level, 
analytic method, and whether or not progress updates should be displayed 
while the function is running.
}
\examples{
\dontrun{
did.count.sim = cps.did.count(nsim = 100, nsubjects = 50, nclusters = 6, c1 = 100, 
                              c2 = 25, sigma_b0 = c(10, 15), sigma_b1 = c(12, 18), 
                              family = 'poisson', analysis = 'poisson', method = 'glmm', 
                              alpha = 0.05, quiet = FALSE, all.sim.data = TRUE)
}

}
\references{
Snjiders, T. & Bosker, R. Multilevel Analysis: an Introduction to Basic and Advanced Multilevel Modelling. London, 1999: Sage.

Elridge, S., Ukoumunne, O. & Carlin, J. The Intra-Cluster Correlation Coefficient in Cluster Randomized Trials: 
A Review of Definitions. International Statistical Review (2009), 77, 3, 378-394. doi: 10.1111/j.1751-5823.2009.00092.x
}
