% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.ma.binary.internal.R
\name{cps.ma.binary.internal}
\alias{cps.ma.binary.internal}
\title{Model fits for simulations for multi-arm designs with dichotomous outcome.}
\usage{
cps.ma.binary.internal(
  nsim = 1000,
  str.nsubjects = NULL,
  probs = NULL,
  sigma_b_sq = NULL,
  alpha = 0.05,
  quiet = FALSE,
  method = "glmm",
  all.sim.data = FALSE,
  seed = NA,
  poor.fit.override = FALSE,
  low.power.override = FALSE,
  timelimitOverride = TRUE,
  tdist = FALSE,
  cores = cores,
  nofit = FALSE,
  optmethod = optmethod,
  return.all.models = FALSE
)
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{str.nsubjects}{Number of subjects per treatment group; accepts a list with one entry per arm.
Each entry is a vector containing the number of subjects per cluster (required).}

\item{probs}{Expected probability of outcome for each arm; accepts a vector of length \code{narms} (required).}

\item{sigma_b_sq}{Between-cluster variance; accepts a vector of length \code{narms} (required).}

\item{alpha}{Significance level; default = 0.05.}

\item{quiet}{When set to FALSE, displays simulation progress and estimated
completion time; default is FALSE.}

\item{method}{Analytical method, either Generalized Linear Mixed Effects
Model (GLMM) or Generalized Estimating Equation (GEE). Accepts c('glmm',
'gee') (required); default = 'glmm'.}

\item{all.sim.data}{Option to output list of all simulated datasets;
default = FALSE.}

\item{seed}{Option to set.seed. Default is NULL.}

\item{poor.fit.override}{Option to override \code{stop()} if more than 25\%
of fits fail to converge.}

\item{low.power.override}{Option to override \code{stop()} if the power
is less than 0.5 after the first 50 simulations and every ten simulations
thereafter. On function execution stop, the actual power is printed in the
stop message. Default = FALSE. When TRUE, this check is ignored and the
calculated power is returned regardless of value.}

\item{timelimitOverride}{Logical. When FALSE, stops execution if the estimated
completion time is more than 2 minutes. Defaults to TRUE.}

\item{tdist}{Logical; use t-distribution instead of normal distribution for
simulation values, default = FALSE.}

\item{cores}{A string ("all") NA, or numeric value indicating the number of cores to be used for parallel computing.
When this option is set to NA, no parallel computing is used.}

\item{nofit}{Option to skip model fitting and analysis and return the simulated data.
Defaults to \code{FALSE}.}

\item{optmethod}{User-specified optimizer method. Accepts \code{bobyqa},
\code{Nelder_Mead}, and optimizers wrapped in the \code{optimx} package.}

\item{return.all.models}{Logical; Returns all of the fitted models and the simulated data.
Defaults to FALSE.}
}
\value{
A list with the following components:
\itemize{
  \item List of length(nsim) containing gee- or glmm-fitted the model summaries.
  \item Compares fitted model to a model for H0 using ML (anova).
  \item List of data frames, each containing:
                  "y" (Simulated response value),
                  "trt" (Indicator for treatment group),
                  "clust" (Indicator for cluster)
  \item A vector of logical values \code{nsim}.
          When a model fails to converge, FALSE, otherwise TRUE.
}
}
\description{
Generally called from \code{cps.ma.binary()}, this function uses iterative
simulations to model significance of treatment effects for cluster-randomized
controlled trials. Users can modify a variety of parameters to suit the
simulations to their desired experimental situation.
}
\details{
This function can be called directly in order to give the user access to the simulated
model fits in addition to the simulated data, the latter of which can also be accessed
here or using the function \code{cps.ma.binary()}. For the power estimates, use
\code{cps.ma.binary()}.

Users (or the wrapper function) must specify the desired number of
simulations, number of subjects per
cluster, number of clusters per treatment arm, group proportions, and
between-cluster variance; significance level, analytic method, progress updates,
and simulated data set output may also be specified.
}
\examples{
\dontrun{

nsubjects.example <- list(c(200,200,200,250), c(150, 200, 200, 210), c(170, 200, 210))
probs.example <- c(0.25, 0.15, 0.3)
sigma_b_sq.example <- c(0.01, 0.01, 0.01)

bin.ma.rct <- cps.ma.binary.internal (nsim = 10,
                            str.nsubjects = nsubjects.example,
                            probs = probs.example,
                            sigma_b_sq = sigma_b_sq.example,
                            alpha = 0.05, all.sim.data = FALSE,
                            poor.fit.override = TRUE,
                            seed = 123, cores="all",
                            optmethod = "Nelder-Mead")
}

}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu}), Alexander R. Bogdan, and Ken Kleinman (\email{ken.kleinman@gmail.com})
}
