% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.ma.normal.internal.R
\name{cps.ma.normal.internal}
\alias{cps.ma.normal.internal}
\title{Runs and returns simulations and model fits for multi-arm 
cluster-randomized trials with continuous outcome.}
\usage{
cps.ma.normal.internal(nsim = 1000, str.nsubjects = NULL,
  means = NULL, sigma_sq = NULL, sigma_b_sq = NULL, alpha = 0.05,
  quiet = FALSE, method = "glmm", all.sim.data = FALSE,
  seed = NULL, cores = "all", poor.fit.override = FALSE,
  tdist = FALSE)
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{str.nsubjects}{Number of subjects per treatment group; accepts a list 
with one entry per arm. Each entry is a vector containing the number of 
subjects per cluster (required).}

\item{means}{Expected absolute treatment effect for each arm; accepts a 
vector of length \code{narms} (required).}

\item{sigma_sq}{Within-cluster variance; accepts a vector of length 
\code{narms} (required).}

\item{sigma_b_sq}{Between-cluster variance; accepts a vector of length 
\code{narms} (required).}

\item{alpha}{Significance level; default = 0.05.}

\item{quiet}{When set to FALSE, displays simulation progress and estimated 
completion time; default is FALSE.}

\item{method}{Analytical method, either Generalized Linear Mixed Effects 
Model (GLMM) or Generalized Estimating Equation (GEE); accepts c('glmm', 
'gee') (required); default = 'glmm'.}

\item{all.sim.data}{Option to output list of all simulated datasets; 
default = FALSE.}

\item{seed}{Option to set.seed, default is NULL.}

\item{cores}{a string or numeric value indicating the number of cores to be 
used for parallel computing. When this option is set to NULL, no parallel 
computing is used.}

\item{poor.fit.override}{Option to override \code{stop()} if more than 25\% 
of fits fail to converge.}

\item{tdist}{Logical; use t-distribution instead of normal distribution 
for simulation values, default = FALSE.}
}
\value{
A list with the following components:
\describe{
  \item{estimates}{List of \code{length(nsim)} containing gee- or glmm-fitted model 
  estimates.}
  \item{model comparisons}{Compares fitted model to a model for H0 using ML 
  (ANOVA).}
  \item{simulated datasets}{Produced when all.sim.data = TRUE, a list of 
  data frames, each containing three columns: "y" (Simulated response value),  
                  "trt" (Indicator for treatment group), and 
                  "clust" (Indicator for cluster).}
  \item{failed.to.converge}{A vector of length \code{nsim} consisting of 1 and 0 
  (when all.sim.data = TRUE) or a string (when all.sim.data = FALSE) 
  indicating the percent of model fits that produced a "singular fit" or "failed to converge" warning message;
          When a model fails to converge, failed.to.converge == 1, otherwise 0.}
}
}
\description{
This function is called within \code{cps.ma.normal()} to generate the 
simulated data and return glmm or gee model fits according to the user's 
specifications. However, this function can also be called independently 
in order to return and examine the simulation model fits rather than the 
power summary returned in \code{cps.ma.normal()}.
}
\details{
Users can modify a variety of parameters to suit the simulations to their
desired experimental situation. Users must specify the desired number 
of simulations, number of subjects per cluster, number of clusters 
per treatment arm, group means, within-cluster variance, and between-cluster 
variance; significance level, analytic method, progress updates, and 
simulated data set output may also be specified.

As this function is not intended for the casual user, specifications to the 
arguments are somewhat less flexible than those supplied to 
\code{cps.ma.normal()}. Note that the str.nsubjects argument has a specific 
structure that is somewhat verbose, but maximizes flexibility in the 
applicable experimental design.
}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu}, Alexander R. Bogdan, and Ken Kleinman (\email{ken.kleinman@gmail.com})
}
