% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.ma.normal.R
\name{cps.ma.normal}
\alias{cps.ma.normal}
\title{Power simulations for cluster-randomized trials: Multi-Arm Designs, Continuous Outcome.}
\usage{
cps.ma.normal(nsim = 1000, nsubjects = NULL, narms = NULL,
  nclusters = NULL, means = NULL, sigma = NULL, sigma_b = NULL,
  alpha = 0.05, quiet = FALSE, ICC = NULL, method = "glmm",
  all.sim.data = FALSE, seed = 123, poor.fit.override = FALSE)
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{nsubjects}{Number of subjects per treatment group; accepts a list with one entry per arm. 
Each entry is a vector containing the number of subjects per cluster (required).}

\item{means}{Expected absolute treatment effect for each arm; accepts a vector of length \code{narms} (required).}

\item{sigma}{Within-cluster variance; accepts a vector of length \code{narms} (required).}

\item{sigma_b}{Between-cluster variance; accepts a vector of length \code{narms} (required).}

\item{alpha}{Significance level; default = 0.05.}

\item{quiet}{When set to FALSE, displays simulation progress and estimated completion time; default is FALSE.}

\item{method}{Analytical method, either Generalized Linear Mixed Effects Model (GLMM) or 
Generalized Estimating Equation (GEE). Accepts c('glmm', 'gee') (required); default = 'glmm'.}

\item{all.sim.data}{Option to output list of all simulated datasets; default = FALSE.}

\item{seed}{Option to set.seed. Default is NULL.}

\item{poor.fit.override}{Option to override \code{stop()} if more than 25% of fits fail to converge or 
power<0.5 after 50 iterations; default = FALSE}
}
\description{
This set of functions utilize iterative simulations to determine 
approximate power for multi-arm cluster-randomized controlled trials. Users 
can modify a variety of parameters to suit the simulations to their
desired experimental situation.
}
\details{
Users must specify the desired number of simulations, number of subjects per 
cluster, number of clusters per treatment arm, group means, two of the following: ICC, within-cluster variance, or 
between-cluster variance. Significance level, analytic method, progress updates, poor/singular fit override,
and simulated data set output may also be specified. This function validates the user's input 
and passes the necessary arguments to \code{cps.ma.normal.internal}, which performs the simulations.
}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu}, Alexander R. Bogdan, and Ken Kleinman (\email{ken.kleinman@gmail.com})
}
