% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMissingVarianceParam.R
\name{createMissingVarianceParam}
\alias{createMissingVarianceParam}
\title{Power simulations for cluster-randomized trials: calculating ICC, 
sigma, and sigma_b inputs}
\usage{
createMissingVarianceParam(ICC = NULL, sigma = NULL, sigma_b = NULL)
}
\arguments{
\item{ICC}{Intra-cluster correlation coefficient; accepts a vector of length \code{narms}}

\item{sigma}{Within-cluster variance; accepts a vector of length \code{narms}.}

\item{sigma_b}{Between-cluster variance; accepts a vector of length \code{narms}.}
}
\value{
A vector of length \code{narms} 
\describe{
  \item{missing.arg}{The argument that the user did not specify is returned as a character 
  vector with the first item being the name of the returned values.}
}
}
\description{
Usually called from within a function, createMissingVarianceParam takes 2 of 3 
arguments (ICC, sigma, or sigma_b) and returns the missing argument. It gives 
an error if one of the inputs is missing or not specified properly. Note that 
it is required that the user specify at least 2 of the arguments.
}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu}, Alexander R. Bogdan, and Ken Kleinman (\email{ken.kleinman@gmail.com})
}
