% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crtpower_2mean.R
\name{crtpower_2mean}
\alias{crtpower_2mean}
\title{Power calculations for simple cluster randomized trials, continuous outcome}
\usage{
crtpower_2mean(alpha = 0.05, power = 0.8, m = NULL, n = NULL,
  nsd = NULL, cv = 0, d = NULL, icc = NULL, varw = NULL,
  varb = NULL, tol = .Machine$double.eps^0.25)
}
\arguments{
\item{alpha}{The level of significance of the test, the probability of a
Type I error.}

\item{power}{The power of the test, 1 minus the probability of a Type II
error.}

\item{m}{The number of clusters per condition. It must be greater than 1.}

\item{n}{The mean of the cluster sizes, or a vector of cluster sizes with
length equal to twice \code{m}.}

\item{nsd}{The standard deviation of the cluster sizes.}

\item{cv}{The coefficient of variation of the cluster sizes. When \code{cv} = 0,
the clusters all have the same size.}

\item{d}{The difference in condition means.}

\item{icc}{The intraclass correlation.}

\item{varw}{The within-cluster variation.}

\item{varb}{The between-cluster variation.}

\item{tol}{Numerical tolerance used in root finding. The default provides
at least four significant digits.}
}
\value{
The computed argument.
}
\description{
Compute the power of a simple cluster randomized trial with a continuous outcome,
or determine parameters to obtain a target power.
}
\details{
Exactly one of \code{alpha}, \code{power}, \code{d}, \code{ICC}, \code{m},
  \code{n}, and \code{cv} must be passed as \code{NULL}. Note that
  \code{alpha}, \code{power}, and \code{cv} have non-\code{NULL}
  defaults, so if those are the parameters of interest they must be
  explicitly passed as \code{NULL}.
}
\section{Note}{

  'uniroot' is used to solve power equation for unknowns, so you may see
  errors from it, notably about inability to bracket the root when
  invalid arguments are given.
}

\section{Authors}{

Jonathan Moyer (\email{jon.moyer@gmail.com})
}

