% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crtpower_2prop.R
\name{crtpower_2prop}
\alias{crtpower_2prop}
\title{Power calculations for simple cluster randomized trials, binary outcome}
\usage{
crtpower_2prop(alpha = 0.05, power = 0.8, m = NULL, n = NULL,
  cv = NULL, p1 = NULL, p2 = NULL, icc = NULL, pooled = FALSE,
  tol = .Machine$double.eps^0.25)
}
\arguments{
\item{alpha}{The level of significance of the test, the probability of a
Type I error.}

\item{power}{The power of the test, 1 minus the probability of a Type II
error.}

\item{m}{The number of clusters per condition. It must be greater than 1.}

\item{n}{The mean of the cluster sizes, or a vector of cluster sizes with
length equal to twice \code{m}.}

\item{cv}{The coefficient of variation of the cluster sizes. When \code{cv} = 0,
the clusters all have the same size.}

\item{p1}{The expected proportion in the treatment group.}

\item{p2}{The proportion in the control group.}

\item{icc}{The intraclass correlation.}

\item{pooled}{Logical indicating if pooled standard error should be used.}

\item{tol}{Numerical tolerance used in root finding. The default provides
at least four significant digits.}
}
\value{
The computed argument.
}
\description{
Compute the power of a simple cluster randomized trial with a binary outcome,
or determine parameters to obtain a target power.
}
\section{Authors}{

Jonathan Moyer (\email{jon.moyer@gmail.com})
}

