% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crtpwr.2meanCA.R
\name{crtpwr.2meanCA}
\alias{crtpwr.2meanCA}
\title{Power calculations for simple cluster randomized trials, continuous outcome with covariate adjustment}
\usage{
crtpwr.2meanCA(alpha = 0.05, power = 0.8, nclusters = NA,
  nsubjects = NA, d = NA, icc = NA, varw = NA, covdf = NA,
  rho_b = NA, tol = .Machine$double.eps^0.25)
}
\arguments{
\item{alpha}{The level of significance of the test, the probability of a
Type I error.}

\item{power}{The power of the test, 1 minus the probability of a Type II
error.}

\item{nclusters}{The number of clusters per condition. It must be greater than 1.}

\item{nsubjects}{The mean of the cluster sizes, or a vector of cluster sizes for one arm.}

\item{d}{The difference in condition means.}

\item{icc}{The intraclass correlation.}

\item{varw}{The within-cluster variation.}

\item{covdf}{The degrees of freedom used by the cluster-level covariates. If \code{covdf} is 0,
the value of \code{rho_b} must be 0 as well.}

\item{rho_b}{The between-cluster residual correlation between the outcome and the covariates.}

\item{tol}{Numerical tolerance used in root finding. The default provides
at least four significant digits.}
}
\value{
The computed argument.
}
\description{
Compute the power of a simple cluster randomized trial with a continuous outcome after adjusting for covariates,
or determine parameters to obtain a target power.
}
\details{
Exactly one of \code{alpha}, \code{power}, \code{nclusters}, \code{nsubjects},
  \code{d}, \code{icc}, \code{varw}, \code{covdf}, and \code{rho_b} must be passed as \code{NA}.
  Note that \code{alpha} and \code{power} have non-\code{NA}
  defaults, so if those are the parameters of interest they must be
  explicitly passed as \code{NA}.
  
If \code{nsubjects} is a vector the values, \code{nclusters} will be recalculated
   using the values in \code{nsubjects}.
}
\section{Note}{

  This function was inspired by work from Stephane Champely (pwr.t.test) and
  Peter Dalgaard (power.t.test). As with those functions, 'uniroot' is used to
  solve power equation for unknowns, so you may see
  errors from it, notably about inability to bracket the root when
  invalid arguments are given.
}

\section{Authors}{

Jonathan Moyer (\email{jon.moyer@gmail.com}), Ken Kleinman (\email{ken.kleinman@gmail.com})
}

\examples{
# Find the number of clusters per condition needed for a trial with alpha = 0.05, 
# power = 0.8, 10 observations per cluster, a difference of 1 unit,  icc = 0.1,
# a variance of 5 units, adjusting for 1 covariate, and a between-cluster correlation of 0.7:
crtpwr.2meanCA(nsubjects=10 ,d=1, icc=.1, varw=5, covdf=1, rho_b=.7)
# 
# The result, showimg nclusters of greater than 12, suggests 13 clusters per condition should be used.
}
\references{
Moerbeek M, Teerenstra S. (2016) Power Analysis of Trials with Multilevel Data. 
  Boca Raton: CRC Press; 2016: 84-87
}
