% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crtpwr.2meanD.ltf.R
\name{crtpwr.2meanD.ltf}
\alias{crtpwr.2meanD.ltf}
\title{Power calculations for difference-in-difference cluster randomized trials accounting for 
loss to and gain at follow-up, continuous outcome}
\usage{
crtpwr.2meanD.ltf(
  alpha = 0.05,
  power = 0.8,
  nclusters = NA,
  nsubjects = NA,
  d = NA,
  icc = NA,
  rho_c = NA,
  rho_s = NA,
  vart = NA,
  ltf = c(0, 0),
  gtf = ltf,
  tol = .Machine$double.eps^0.25
)
}
\arguments{
\item{alpha}{The level of significance of the test, the probability of a
Type I error.}

\item{power}{The power of the test, 1 minus the probability of a Type II
error.}

\item{nclusters}{The number of clusters per condition at baseline. It must be greater than 1.}

\item{nsubjects}{The number of subjects per cluster at baseline.}

\item{d}{The difference in mean change between conditions (i.e. "difference-in-difference").}

\item{icc}{The intraclass correlation.}

\item{rho_c}{The correlation between baseline and post-test outcomes at the
cluster level. This value can be used in both cross-sectional and cohort
designs. If this quantity is unknown, a value of 0 is a conservative estimate.}

\item{rho_s}{The correlation between baseline and post-test outcomes at the
subject level assuming the a cohort design with no loss or gain to follow-up. Entering a value
of 0 for this parameter is equivalent to a cross-sectional design and automatically sets \code{ltf_0}, \code{ltf_1},
\code{gtf_0}, and \code{gtf_1} all to 1.}

\item{vart}{The total variation of the outcome (the sum of within- and between-cluster variation).}

\item{ltf}{A length 2 vector indicating the proportion lost to follow-up per cluster in terms of \code{nsubjects}
in the control and treatment arms. The first element corresponds to the proportion lost in the control
group, while the second element correponds to the proportion lost in the treatment group. Entering a scalar
for \code{ltf} will set both elements of the vector to that value. If \code{rho_s} is 0 then both elements 
will be set to 1.}

\item{gtf}{A length 2 vector indicating the proportion gained at follow-up per cluster in terms of \code{nsubjects}
in the control and treatment arms. The first element corresponds to the proportion gained in the control
group, while the second element correponds to the proportion gained in the treatment group. By default, \code{gtf}
is set to \code{ltf}. Entering a scalar for \code{gtf} will set both elements of the vector to that value. 
If \code{rho_s} is 0 then both elements will be set to 1.}

\item{tol}{Numerical tolerance used in root finding. The default provides
at least four significant digits.}
}
\value{
The computed argument.
}
\description{
Compute the power of a difference-in-difference cluster 
randomized trial design with a continuous outcome,
accounting for potential loss to or gain at follow-up, or 
determine parameters to obtain a target power.
}
\details{
Exactly one of \code{alpha}, \code{power}, \code{nclusters}, \code{nsubjects},
  \code{d}, \code{icc}, \code{rho_c}, \code{rho_s}, and \code{vart}
  must be passed as \code{NA}. Note that \code{alpha} and\code{power}
  have non-\code{NA} defaults, so if those are the parameters of 
  interest they must be explicitly passed as \code{NA}.
}
\section{Note}{

  This function was inspired by work from Stephane Champely (pwr.t.test) and
  Peter Dalgaard (power.t.test). As with those functions, 'uniroot' is used to
  solve power equation for unknowns, so you may see
  errors from it, notably about inability to bracket the root when
  invalid arguments are given.
}

\section{Authors}{



Jonathan Moyer (\email{jon.moyer@gmail.com}), Ken Kleinman (\email{ken.kleinman@gmail.com})
}

\examples{
# Find the number of clusters per condition needed for a trial with alpha = 0.05, 
# power = 0.80, nsubjects = 20, d = 0.50 units, icc = 0.05, rho_c = 0.50, rho_s = 0.70,
# and vart = 1 square unit if 50 percent of subjects in each cluster are lost to follow-up 
# and replaced.
crtpwr.2meanD.ltf(nsubjects = 100 , d = 0.5, icc = 0.05, 
  rho_c = 0.50, rho_s = 0.70, vart = 1,ltf=0.5)
# 
# The result, nclusters = 8.099772, suggests 9 clusters per condition should be used.

}
\references{
Rutterford C, Copas A, Eldridge S. (2015) Methods for sample size
  determination in cluster randomized trials. Int J Epidemiol. 44(3):1051-1067.

Teerenstra S, Eldridge S, Graff M, de Hoop E, Borm, GF. (2012) A simple
  sample size formula for analysis of covariance in cluster randomized trials.
  Statist Med. 31:2169-2178
}
