% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crtpwr.2propM.R
\name{crtpwr.2propM}
\alias{crtpwr.2propM}
\title{Power calculations for simple cluster randomized trials with matching, binary outcome}
\usage{
crtpwr.2propM(alpha = 0.05, power = 0.8, nclusters = NA, nsubjects = NA,
  p1 = NA, p2 = NA, cvm = 0, p1inc = TRUE,
  tol = .Machine$double.eps^0.25)
}
\arguments{
\item{alpha}{The level of significance of the test, the probability of a
Type I error.}

\item{power}{The power of the test, 1 minus the probability of a Type II
error.}

\item{nclusters}{The number of clusters per condition. It must be greater than 1.}

\item{nsubjects}{The mean of the cluster sizes.}

\item{p1}{The expected proportion in the treatment group.}

\item{p2}{The proportion in the control group.}

\item{cvm}{The coefficient of variation in the outcome within matched clusters.}

\item{p1inc}{Logical indicating if p1 is expected to be greater than p2.}

\item{tol}{Numerical tolerance used in root finding. The default provides
at least four significant digits.}
}
\value{
The computed argument.
}
\description{
Compute the power of a simple cluster randomized trial with a binary outcome with cluster-level matching,
or determine parameters to obtain a target power.
}
\section{Authors}{

Jonathan Moyer (\email{jon.moyer@gmail.com}), Ken Kleinman (\email{ken.kleinman@gmail.com})
}

\section{Note}{

  This function was inspired by work from Stephane Champely (pwr.t.test) and
  Peter Dalgaard (power.t.test). As with those functions, 'uniroot' is used to
  solve power equation for unknowns, so you may see
  errors from it, notably about inability to bracket the root when
  invalid arguments are given.
}

\examples{
# Find the number of clusters per condition needed for a trial with alpha = 0.05, 
# power = 0.80, 80 observations per probability
# in condition 1 of 0.80 and condition 2 of 0.60, within-pair coefficient of 
# variation equal to 0.2.
crtpwr.2propM(nsubjects=80 ,p1=0.80, p2=0.60, cvm = 0.2)
# 
# The result, showimg nclusters of greater than 10, suggests 11 clusters per 
# condition should be used.

}
\references{
Hayes JR, Moulton LH. Cluster Randomized Trials. Boca Raton, FL: CRC Press; 2009.
}
