% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crtpwr.nmean.R
\name{crtpwr.nmean}
\alias{crtpwr.nmean}
\title{Power calculations for multi-arm cluster randomized trials, continuous outcome}
\usage{
crtpwr.nmean(alpha = 0.05, power = 0.8, narms = NA, nclusters = NA,
  nsubjects = NA, vara = NA, varc = NA, vare = NA,
  tol = .Machine$double.eps^0.25)
}
\arguments{
\item{alpha}{The level of significance of the test, the probability of a
Type I error.}

\item{power}{The power of the test, 1 minus the probability of a Type II
error.}

\item{narms}{The number of independent arms (conditions). It must be greater than 2.}

\item{nclusters}{The number of clusters per arm. It must be greater than 1.}

\item{nsubjects}{The cluster size.}

\item{vara}{The between-arm variance.}

\item{varc}{The between-cluster variance.}

\item{vare}{The within-cluster variance.}

\item{tol}{Numerical tolerance used in root finding. The default provides
at least four significant digits.}
}
\value{
The computed argument.
}
\description{
Compute the power of a multi-arm cluster randomized trial with a continuous outcome,
or determine parameters to obtain a target power.
}
\details{
Exactly one of \code{alpha}, \code{power}, \code{narms}, \code{nclusters},
  \code{nsubjects}, \code{vara}, \code{varc}, and \code{vare}  must be passed as \code{NA}.
  Note that \code{alpha} and \code{power} have non-\code{NA}
  defaults, so if those are the parameters of interest they must be
  explicitly passed as \code{NA}.
  
Assuming a balanced design, the between-arm variance \eqn{\sigma_a^2} (corresponding to
  the function argument \code{vara}) can be estimated using the formula: 
  
  \deqn{\sigma_a^2 = \sum\limits_{i=1}^{n_a}(\mu_i - \mu)^2/(n_a-1)}
  
  where \eqn{n_a} is the number of arms, \eqn{\mu_i} is the estimate of the \eqn{i}-th arm
  mean, and \eqn{\mu} is the estimate of the overall mean of the outcome. This 
  variance can be computed in R using the \code{var} function and a vector of arm means.
  For example, suppose the estimated means for a four arm trial were 40, 50, 50, and 55. Then the
  estimate of the between-arm variance could be computed with \code{var(c(40,50,50,55))}, 
  yielding a value of about 39.58.
}
\section{Note}{

  This function was inspired by work from Stephane Champely (pwr.t.test),
  Peter Dalgaard (power.t.test), and Claus Ekstrom (power.anova.test). As 
  with those functions, 'uniroot' is used to solve power equation for
  unknowns, so you may see errors from it, notably about inability to
  bracket the root when invalid arguments are given.
}

\section{Authors}{

Jonathan Moyer (\email{jon.moyer@gmail.com}), Ken Kleinman (\email{ken.kleinman@gmail.com})
}

\examples{
# Find the number of clusters per condition needed for a 4-arm trial with alpha = .05, 
# power = 0.8, 10 observations per cluster, between-arm variance of 50,
# between-cluster variance of 10, and within-cluster variance of 60. 
crtpwr.nmean(narms=4,nsubjects=10,vara=50,varc=10,vare=60)
# 
# The result, showimg nclusters of greater than 5, suggests 6 clusters per 
# condition should be used.
}
\references{
Murray DM. Design and Analysis of Group-Randomized Trials. New York,
  NY: Oxford University Press; 1998.
}
