% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.wholenumber.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Power simulations for cluster-randomized trials: check if whole number.}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{The numeric value to be tested.}
}
\value{
Logical.
}
\description{
Power simulations for cluster-randomized trials: check if whole number.
}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu} and Ken Kleinman (\email{ken.kleinman@gmail.com})
}
