% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizerSearch.R
\name{optimizerSearch}
\alias{optimizerSearch}
\title{Power simulations for cluster-randomized trials: find an optimizer.}
\usage{
optimizerSearch(model)
}
\arguments{
\item{model}{a lme4 or glmer model}
}
\value{
A string
\describe{
  \item{goodopt}{The name of the first optimizer tested that showed convergence.}
}
}
\description{
This function can be called by some of the clusterPower functions to search 
for an optimizer that does not produce convergence errors.
}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu}, and Joshua Nugent.
}
