% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_map_helper.R
\name{package_map_helper}
\alias{package_map_helper}
\title{Look up which internal functions are called by exported functions.}
\usage{
package_map_helper(packageName = "clusterPower")
}
\arguments{
\item{packageName}{The name of the package in quotes. Defaults to "clusterPower".}
}
\value{
List of internal functions and the line numbers in which they appear 
inside external functions.
}
\description{
Look up which internal functions are called by exported functions.
}
\author{
Alexandria Sakrejda (\email{acbro0@umass.edu})
}
