% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.method.crtpwr.R
\name{print.crtpwr}
\alias{print.crtpwr}
\title{A print method for crtpwr objects}
\usage{
\method{print}{crtpwr}(x, ...)
}
\arguments{
\item{x}{A crtpwr object}

\item{...}{Other arguments passed to the function}
}
\description{
A print method for crtpwr objects
}
