% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run a Shiny app for power analysis functions}
\usage{
runExample(appname = "main")
}
\arguments{
\item{appname}{which app should be launched. Choices are either "main" or "analytic".
Default is "main".}
}
\description{
This function runs the clusterPower Shiny apps.
}
\author{
Jonathan Moyer (\email{jon.moyer@gmail.com})

Alexandria Sakrejda (\email{acbro0@umass.edu})
}
