% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.method.crtpwr.2arm.R
\name{summary,crtpwr-method}
\alias{summary,crtpwr-method}
\title{A summary method for 2-arm simulation crtpwr objects (crtpwr)}
\usage{
\S4method{summary}{crtpwr}(object, ...)
}
\arguments{
\item{object}{A crtpwr object}

\item{...}{other arguments passed to the summary method}
}
\description{
A summary method for 2-arm simulation crtpwr objects (crtpwr)
}
