% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.method.crtpwr.ma.R
\name{summary,crtpwr.ma-method}
\alias{summary,crtpwr.ma-method}
\title{A summary method for multi-arm crtpwr objects (crtpwr.ma)}
\usage{
\S4method{summary}{crtpwr.ma}(object, ...)
}
\arguments{
\item{object}{A crtpwr.ma object}

\item{...}{other arguments passed to the summary method}
}
\description{
A summary method for multi-arm crtpwr objects (crtpwr.ma)
}
