% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type1ErrTest.R
\name{type1ErrTest}
\alias{type1ErrTest}
\title{Power simulations for cluster-randomized trials: trigger a conservative estimate warning.}
\usage{
type1ErrTest(sigma_sq_, sigma_b_sq_, nsubjects_)
}
\arguments{
\item{sigma_sq_}{User-selected sigma_sq value.}

\item{sigma_b_sq_}{User-selected sigma_b_sq value.}

\item{nsubjects_}{User-selected number of observations.}
}
\value{
If triggered, a warning. Otherwise silent.
\describe{
  \item{warning}{With the parameters you entered, the power estimate may be conservative.}
}
}
\description{
ClusterPower uses maximum likelihood inference, which depends on 
asymptotic results. These results hove known limitations in finite 
settings, such as when there are small numbers of clusters or small 
numbers of observation per cluster.
}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu} and Ken Kleinman (\email{ken.kleinman@gmail.com})
}
