% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateVariance.R
\name{validateVariance}
\alias{validateVariance}
\title{Power simulations for cluster-randomized trials: validating ICC, 
sigma, and sigma_b inputs}
\usage{
validateVariance(difference = means, alpha = alpha, ICC = ICC,
  sigma = sigma, sigma_b = sigma_b, ICC2 = ICC, sigma2 = sigma,
  sigma_b2 = sigma_b, method = method, quiet = quiet,
  all.sim.data = all.sim.data, poor.fit.override = poor.fit.override)
}
\arguments{
\item{difference}{Expected absolute treatment effect; accepts numeric.}

\item{alpha}{Significance level; default = 0.05.}

\item{ICC}{Intra-cluster correlation coefficient; accepts a vector of length \code{narms}
with values between 0 - 1.}

\item{sigma}{Within-cluster variance; accepts numeric}

\item{sigma_b}{Between-cluster variance; accepts numeric
If clusters differ between treatment groups, at least 2 of the following 
must be specified:}

\item{ICC2}{Intra-cluster correlation coefficient for clusters in TREATMENT group}

\item{sigma2}{Within-cluster variance for clusters in TREATMENT group}

\item{sigma_b2}{Between-cluster variance for clusters in TREATMENT group}

\item{method}{Analytical method, either Generalized Linear Mixed Effects Model (GLMM) or 
Generalized Estimating Equation (GEE). Accepts c('glmm', 'gee') (required); default = 'glmm'.}

\item{quiet}{Logical. When set to FALSE, displays simulation progress and estimated completion time; default is FALSE.}

\item{all.sim.data}{Logical. Option to output list of all simulated datasets; default = FALSE.}

\item{ICC}{Intra-cluster correlation coefficient; accepts a value between 0 - 1}
}
\value{
A vector of length \code{narms} 
\describe{
  \item{missing.arg}{The argument that the user did not specify is returned as a character 
  vector with the first item being the name of the returned values.}
}
}
\description{
Usually called from within a function, validateVariance takes variance parameters
and validates those inputs. This function gives an error if one of the inputs 
is missing or not specified properly.
}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu}, Alexander R. Bogdan, and Ken Kleinman (\email{ken.kleinman@gmail.com})
}
