% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpa.sw.binary.R
\name{cpa.sw.binary}
\alias{cpa.sw.binary}
\title{Power simulations for cluster-randomized trials: Stepped Wedge Design, Binary Outcome}
\usage{
cpa.sw.binary(
  nclusters = 12,
  ntimes = 3,
  nsubjects = 40,
  d = -10,
  ICC = 0.01,
  beta = -2,
  mu = 0.18,
  tol = 1e-05,
  GQ = 100
)
}
\arguments{
\item{nclusters}{Number of clusters; accepts non-negative integer scalar (required).}

\item{nsubjects}{Number of subjects per cluster; accepts either a scalar (equal cluster sizes) 
or a vector of length \code{nclusters} (user-defined size for each cluster) (required).}

\item{steps}{Number of crossover steps; a baseline step (all clusters in non-treatment group) is assumed. 
Accepts positive scalar (indicating the total number of steps; clusters per step is obtained by 
\code{nclusters / steps}) or a vector of non-negative integers corresponding either to the number 
of clusters to be crossed over at each time point (e.g c(2,4,4,2); nclusters = 10) or the cumulative 
number of clusters crossed over by a given time point (e.g. c(2,4,8,10); nclusters = 10) (required).}

\item{alpha}{Significance level. Default = 0.05.}

\item{quiet}{When set to FALSE, displays simulation progress and estimated completion time; default is FALSE.}
}
\value{
The estimated power.
}
\description{
This function uses a modified Cox method to determine power for stepped 
wedge cluster-randomized controlled trials. Users can modify a variety of 
parameters to suit their desired experimental situation.
}
\details{
The stepped wedge trial design is a type of cross-over
design in which clusters change treatments in waves. Initially all the 
clusters recieve the same standard treatment, and at the end of the trial all
of the clusters will be recieving the treatment of interest. More than one 
cluster can change treatments in a wave, but the order in which clusters 
change treatments is randomly determined. The outcome of interest is assessed 
in each cluster during each wave.

Users must specify the number of subjects per cluster, number of clusters 
per treatment arm, the number of time steps, the baseline effect, the 
expected treatment effect, expected absolute difference between treatment 
arms, ICC, and significance level.
}
\examples{
\dontrun{
holder <- cpa.sw.binary()
}

}
\references{
Zhou X, Liao X, Kunz LM, Normand ST, Wang M, Spiegelman D. A maximum 
likelihood approach to power calculations for stepped wedge designs of binary 
outcomes. Biostatistics. 2020 Jan 1;21(1):102-121. doi: 10.1093/biostatistics/kxy031
}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu})

Ken Kleinman (\email{ken.kleinman@gmail.com})
}
