% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.binary.R
\name{cps.binary}
\alias{cps.binary}
\title{Power simulations for cluster-randomized trials: Simple Designs, Binary Outcome.}
\usage{
cps.binary(nsim = NULL, m = NULL, n = NULL, p.diff = NULL, p1 = NULL,
  p2 = NULL, or1 = NULL, or2 = NULL, or.diff = NULL, sigma_b = NULL,
  sigma_b2 = NULL, alpha = 0.05, method = "glmm", quiet = FALSE)
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{m}{Number of clusters per treatment group; accepts integer (required).
At least 2 of the following 3 arguments must be specified when using expected probabilities:}

\item{n}{Number of subjects per cluster; accepts integer (required).}

\item{p.diff}{Expected difference in probability of outcome between groups, defined as p.diff = p1 - p2
At least 2 of the following 3 arguments must be specified when using expected odds ratios:}

\item{p1}{Expected probability of outcome in non-treatment group}

\item{p2}{Expected probability of outcome in treatment group}

\item{or1}{Expected odds ratio for outcome in non-treatment group}

\item{or2}{Expected odds ratio for outcome in treatment group}

\item{or.diff}{Expected difference in odds ratio for outcome between groups, defined as or.diff = or1 - or2}

\item{sigma_b}{Between-cluster variance; if sigma_b2 is not specified, 
between cluster variances are assumed to be equal for both groups. Accepts numeric.
If between cluster variances differ between treatment groups, sigma_b2 must also be specified:}

\item{sigma_b2}{Between-cluster variance for clusters in TREATMENT group}

\item{alpha}{Significance level. Default = 0.05}

\item{method}{Analytical method, either Generalized Linear Mixed Effects Model (GLMM) or Generalized Estimating Equation (GEE). Accepts c('glmm', 'gee') (required); default = 'glmm'.}

\item{quiet}{When set to FALSE, displays simulation progress and estimated completion time. Default is FALSE.}
}
\description{
This function utilizes iterative simulations to determine 
approximate power for cluster-randomized controlled trials. Users 
can modify a variety of parameters to suit the simulations to their
desired experimental situation.
}
\details{
Runs the power simulation for binary outcomes.

Users must specify the desired number of simulations, number of subjects per 
cluster, number of clusters per treatment arm, two of the following three terms: 
expected probability of outcome in non-treatment group, expected probability of 
outcome in treatment group, expected difference in probabilities between groups
; significance level, analytic method, and whether or not progress updates should 
be displayed while the function is running.
}
\examples{
\dontrun{
my.binary.sim = cps.binary(nsim = 100, n = 50, m = 6, p1 = 0.4, p2 = 0.2, sigma_b = 100,
                    alpha = 0.05, method = 'glmm', quiet = FALSE)
}

}
\author{
Alexander R. Bogdan
}
