% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.count.R
\name{cps.count}
\alias{cps.count}
\title{Power simulations for cluster-randomized trials: Simple Designs, Count Outcome.}
\usage{
cps.count(nsim = NULL, nsubjects = NULL, nclusters = NULL,
  c1 = NULL, c2 = NULL, c.diff = NULL, sigma_b = NULL,
  sigma_b2 = NULL, family = "poisson", analysis = "poisson",
  method = "glmm", alpha = 0.05, quiet = FALSE,
  all.sim.data = FALSE)
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{nsubjects}{Number of subjects per cluster; accepts integer (required).}

\item{nclusters}{Number of clusters per treatment group; accepts integer (required).
At least 2 of the following 3 arguments must be specified:}

\item{c1}{Expected outcome count in non-treatment group}

\item{c2}{Expected outcome count in treatment group}

\item{c.diff}{Expected difference in outcome count between groups, defined as c.diff = c1 - c2}

\item{sigma_b}{Between-cluster variance; if sigma_b2 is not specified, 
between cluster variances are assumed to be equal between groups. Accepts numeric
If between cluster variances differ between treatment groups, the following must also be specified:}

\item{sigma_b2}{Between-cluster variance for clusters in TREATMENT group}

\item{family}{Distribution from which responses are simulated. Accepts Poisson ('poisson') or negative binomial ('neg.binom') (required); default = 'poisson'}

\item{analysis}{Family used for regression; currently only applicable for GLMM. Accepts c('poisson', 'neg.binom') (required); default = 'poisson'}

\item{method}{Analytical method, either Generalized Linear Mixed Effects Model (GLMM) or Generalized Estimating Equation (GEE). Accepts c('glmm', 'gee') (required); default = 'glmm'}

\item{alpha}{Significance level. Default = 0.05.}

\item{quiet}{When set to FALSE, displays simulation progress and estimated completion time. Default = FALSE.}

\item{all.sim.data}{Option to output list of all simulated datasets. Default = FALSE}
}
\description{
This function utilizes iterative simulations to determine 
approximate power for cluster-randomized controlled trials. Users 
can modify a variety of parameters to suit the simulations to their
desired experimental situation.
}
\details{
Runs the power simulation for count outcomes.

Users must specify the desired number of simulations, number of subjects per 
cluster, number of clusters per treatment arm, between-cluster variance, 
two of the following: expected count in non-treatment group, expected count 
in treatment group, difference in counts between groups; significance level, 
analytic method, and whether or not progress updates should be displayed 
while the function is running.
}
\examples{
\dontrun{
count.sim = cps.count(nsim = 100, nsubjects = 50, nclusters = 6, c1 = 100,
                      c2 = 25, sigma_b = 100, family = 'poisson', analysis = 'poisson',
                      method = 'glmm', alpha = 0.05, quiet = FALSE, all.sim.data = TRUE)
}

}
\author{
Alexander R. Bogdan
}
