% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.count.R
\name{cps.count}
\alias{cps.count}
\title{Power simulations for cluster-randomized trials: Simple Designs, Count Outcome.}
\usage{
cps.count(
  nsim = NULL,
  nsubjects = NULL,
  nclusters = NULL,
  c1 = NULL,
  c2 = NULL,
  c.diff = NULL,
  sigma_b_sq = NULL,
  sigma_b_sq2 = NULL,
  family = "poisson",
  analysis = "poisson",
  method = "glmm",
  alpha = 0.05,
  quiet = FALSE,
  all.sim.data = FALSE,
  irgtt = FALSE,
  seed = NA,
  optimizer = "L-BFGS-B"
)
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{nsubjects}{Number of subjects per cluster; accepts a single integer or a vector of 2 integers (if nsubjects differs between arms) (required).}

\item{nclusters}{Number of clusters per treatment group; accepts a single integer or a vector of 2 integers (if nsubjects differs between arms) (required).
At least 2 of the following 3 arguments must be specified:}

\item{c1}{Expected outcome count in non-treatment group}

\item{c2}{Expected outcome count in treatment group}

\item{c.diff}{Expected difference in outcome count between groups, defined as c.diff = c1 - c2}

\item{sigma_b_sq}{Between-cluster variance; if sigma_b_sq2 is not specified, 
between cluster variances are assumed to be equal between groups. Accepts numeric
If between cluster variances differ between treatment groups, the following must also be specified:}

\item{sigma_b_sq2}{Between-cluster variance for clusters in TREATMENT group}

\item{family}{Distribution from which responses are simulated. Accepts Poisson ('poisson') or negative binomial ('neg.binom') (required); default = 'poisson'}

\item{analysis}{Family used for regression; currently only applicable for GLMM. Accepts c('poisson', 'neg.binom') (required); default = 'poisson'}

\item{method}{Analytical method, either Generalized Linear Mixed Effects Model (GLMM) or Generalized Estimating Equation (GEE). Accepts c('glmm', 'gee') (required); default = 'glmm'}

\item{alpha}{Significance level. Default = 0.05.}

\item{quiet}{When set to FALSE, displays simulation progress and estimated completion time. Default = FALSE.}

\item{all.sim.data}{Option to output list of all simulated datasets. Default = FALSE}

\item{optimizer}{Option to fit with a different optimizer (using the package \code{optimx}). Defaults to L-BFGS-B.}
}
\value{
A list with the following components
\itemize{
  \item Character string indicating total number of simulations, distribution of 
  simulated data, and regression family
  \item Number of simulations
  \item Data frame with columns "Power" (Estimated statistical power), 
               "lower.95.ci" (Lower 95% confidence interval bound), 
               "upper.95.ci" (Upper 95% confidence interval bound).
               Note that non-convergent models are returned for review, 
               but not included in this calculation.
  \item Analytic method used for power estimation
  \item Data frame containing families for distribution and analysis of simulated data
  \item Significance level
  \item Vector containing user-defined cluster sizes
  \item Vector containing user-defined number of clusters
  \item Data frame reporting between-cluster variances for Treatment/Non-Treatment groups
  \item Vector containing expected counts and risk ratios based on user inputs
  \item Data frame with columns: 
                  "Estimate" (Estimate of treatment effect for a given simulation), 
                  "Std.err" (Standard error for treatment effect estimate), 
                  "Test.statistic" (z-value (for GLMM) or Wald statistic (for GEE)), 
                  "p.value",
                  "sig.val" (Is p-value less than alpha?)
  \item List of data frames, each containing: 
                  "y" (Simulated response value), 
                  "trt" (Indicator for treatment group), 
                  "clust" (Indicator for cluster)
}
}
\description{
This function utilizes iterative simulations to determine 
approximate power for cluster-randomized controlled trials. Users 
can modify a variety of parameters to suit the simulations to their
desired experimental situation.
}
\details{
Runs the power simulation for count outcomes.

Users must specify the desired number of simulations, number of subjects per 
cluster, number of clusters per treatment arm, between-cluster variance, 
two of the following: expected count in non-treatment group, expected count 
in treatment group, difference in counts between groups; significance level, 
analytic method, and whether or not progress updates should be displayed 
while the function is running.

Non-convergent models are not included in the calculation of exact confidence 
intervals.
}
\examples{
\dontrun{
count.sim = cps.count(nsim = 100, nsubjects = 20, nclusters = 10, 
                      c1 = 20, c2 = 30, sigma_b_sq = 0.1, 
                      family = 'poisson', analysis = 'poisson',
                      method = 'glmm', alpha = 0.05, quiet = FALSE, 
                      all.sim.data = FALSE, seed = 123, optimizer = "L-BFGS-B")
}

}
\author{
Alexander R. Bogdan

Alexandria C. Sakrejda (\email{acbro0@umass.edu}

Ken Kleinman (\email{ken.kleinman@gmail.com})
}
