% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.did.count.R
\name{cps.did.count}
\alias{cps.did.count}
\title{Power simulations for cluster-randomized trials: Difference in Difference, Count Outcome.}
\usage{
cps.did.count(nsim = NULL, nsubjects = NULL, nclusters = NULL,
  c1 = NULL, c2 = NULL, c.diff = NULL, sigma_b0 = NULL, sigma_b1 = 0,
  family = "poisson", analysis = "poisson", method = "glmm",
  alpha = 0.05, quiet = FALSE, all.sim.data = FALSE)
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{nsubjects}{Number of subjects per cluster; accepts integer (required).}

\item{nclusters}{Number of clusters per treatment group; accepts integer (required).
At least 2 of the following 3 arguments must be specified:}

\item{c1}{Expected outcome count in non-treatment group}

\item{c2}{Expected outcome count in treatment group}

\item{c.diff}{Expected difference in outcome count between groups, defined as c.diff = c1 - c2}

\item{sigma_b0}{Pre-treatment (time == 0) between-cluster variance; accepts numeric scalar (indicating equal 
between-cluster variances for both treatment groups) or a vector of length 2 specifying treatment-specific 
between-cluster variances}

\item{sigma_b1}{Post-treatment (time == 1) between-cluster variance; accepts numeric scalar (indicating equal 
between-cluster variances for both treatment groups) or a vector of length 2 specifying treatment-specific 
between-cluster variances. For data simulation, SIGMA_B1 is added to SIGMA_B0, such that if SIGMA_B0 = 5 
and SIGMA_B1 = 2, the between-cluster variance at time == 1 equals 7. Default = 0.}

\item{family}{Distribution from which responses are simulated. Accepts Poisson ('poisson') or negative binomial ('neg.binom') (required); default = 'poisson'}

\item{analysis}{Family used for regression; currently only applicable for GLMM. Accepts c('poisson', 'neg.binom') (required); default = 'poisson'}

\item{method}{Analytical method, either Generalized Linear Mixed Effects Model (GLMM) or Generalized Estimating Equation (GEE). Accepts c('glmm', 'gee') (required); default = 'glmm'}

\item{alpha}{Significance level. Default = 0.05.}

\item{quiet}{When set to FALSE, displays simulation progress and estimated completion time. Default = FALSE.}

\item{all.sim.data}{Option to output list of all simulated datasets. Default = FALSE}
}
\description{
This function utilizes iterative simulations to determine 
approximate power for cluster-randomized controlled trials. Users 
can modify a variety of parameters to suit the simulations to their
desired experimental situation.
}
\details{
Runs power simulations for difference in difference cluster randomized control trials using count outcomes

Users must specify the desired number of simulations, number of subjects per 
cluster, number of clusters per treatment arm, between-cluster variance, 
two of the following: expected count in non-treatment group, expected count 
in treatment group, difference in counts between groups; significance level, 
analytic method, and whether or not progress updates should be displayed 
while the function is running.
}
\examples{
\dontrun{
did.count.sim = cps.did.count(nsim = 100, nsubjects = 50, nclusters = 6, c1 = 100, 
                              c2 = 25, sigma_b0 = c(10, 15), sigma_b1 = c(12, 18), 
                              family = 'poisson', analysis = 'poisson', method = 'glmm', 
                              alpha = 0.05, quiet = FALSE, all.sim.data = TRUE)
}

}
\author{
Alexander R. Bogdan
}
