% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.did.normal.R
\name{cps.did.normal}
\alias{cps.did.normal}
\title{Power simulations for cluster-randomized trials: Difference in Difference Design, Continuous Outcome.}
\usage{
cps.did.normal(nsim = NULL, nsubjects = NULL, nclusters = NULL,
  difference = NULL, sigma = NULL, sigma_b0 = NULL, sigma_b1 = 0,
  alpha = 0.05, method = "glmm", quiet = FALSE,
  all.sim.data = FALSE)
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{nsubjects}{Number of subjects per treatment group; accepts either a scalar (equal cluster sizes, both groups), 
a vector of length two (equal cluster sizes within groups), or a vector of length \code{sum(nclusters)} 
(unequal cluster sizes within groups) (required).}

\item{nclusters}{Number of clusters per group; accepts integer scalar or vector of length 2 for unequal number 
of clusters per treatment group (required)}

\item{difference}{Expected absolute difference in treatment effect between time points; accepts numeric (required).}

\item{sigma}{Within-cluster variance; accepts numeric scalar (indicating equal within-cluster variances for both 
treatment groups at both time points) or vector of length 4 specifying within-cluster variance for each treatment 
group at each time point.}

\item{sigma_b0}{Pre-treatment (time == 0) between-cluster variance; accepts numeric scalar (indicating equal 
between-cluster variances for both treatment groups) or a vector of length 2 specifying treatment-specific 
between-cluster variances}

\item{sigma_b1}{Post-treatment (time == 1) between-cluster variance; accepts numeric scalar (indicating equal 
between-cluster variances for both treatment groups) or a vector of length 2 specifying treatment-specific 
between-cluster variances. For data simulation, SIGMA_B1 is added to SIGMA_B0, such that if SIGMA_B0 = 5 
and SIGMA_B1 = 2, the between-cluster variance at time == 1 equals 7. Default = 0.}

\item{alpha}{Significance level. Default = 0.05.}

\item{method}{Analytical method, either Generalized Linear Mixed Effects Model (GLMM) or 
Generalized Estimating Equation (GEE). Accepts c('glmm', 'gee') (required); default = 'glmm'.}

\item{quiet}{When set to FALSE, displays simulation progress and estimated completion time; default is FALSE.}

\item{all.sim.data}{Option to output list of all simulated datasets; default = FALSE.}
}
\description{
This set of functions utilize iterative simulations to determine 
approximate power for difference in difference cluster-randomized controlled trials. Users 
can modify a variety of parameters to suit the simulations to their
desired experimental situation.
}
\details{
Runs the power simulation for difference in difference (DID) cluster-randomized controlled trial.

Users must specify the desired number of simulations, number of subjects per 
cluster, number of clusters per treatment arm, expected absolute difference 
between treatment arms, two of the following: ICC, within-cluster variance, or 
between-cluster variance; significance level, analytic method, progress updates, 
and simulated data set output may also be specified.
}
\examples{
\dontrun{
normal.did.rct = cps.did.normal(nsim = 100, nsubjects = 150, nclusters = 6, 
                                difference = .48, sigma = 1, alpha = 0.05, 
                                sigma_b0 = .1, method = 'glmm', quiet = FALSE, 
                                all.sim.data = FALSE)
}

}
\author{
Alexander R. Bogdan
}
