% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.irgtt.normal.R
\name{cps.irgtt.normal}
\alias{cps.irgtt.normal}
\title{Simulation-based power estimation for continuous outcome individually 
randomized group treatment trials.}
\usage{
cps.irgtt.normal(
  nsim = NULL,
  nsubjects = NULL,
  nclusters = NULL,
  difference = NULL,
  ICC = NULL,
  sigma_sq = NULL,
  sigma_b_sq = 0,
  ICC2 = ICC,
  sigma_sq2 = sigma_sq,
  sigma_b_sq2 = 0,
  alpha = 0.05,
  method = "glmm",
  quiet = FALSE,
  all.sim.data = FALSE,
  seed = NA,
  poor.fit.override = FALSE
)
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{nsubjects}{Number of subjects per treatment group; accepts either a scalar (equal cluster sizes, both groups), 
a vector of length two (equal cluster sizes within groups), or a vector of length \code{sum(nclusters)} 
(unequal cluster sizes within groups) (required).}

\item{nclusters}{Number of clusters in the clustered group; accepts a scalar (required)}

\item{difference}{Expected absolute treatment effect; accepts numeric (required).}

\item{ICC}{Intra-cluster correlation coefficient; accepts a value between 0 - 1}

\item{sigma_sq}{Within-cluster variance; accepts numeric}

\item{sigma_b_sq}{Between-cluster variance; defaults to 0. Accepts numeric.
If clusters differ between treatment groups, at least 1 of the following 
must be specified: ICC2, sigma_sq2.}

\item{ICC2}{Intra-cluster correlation coefficient for clusters in TREATMENT group}

\item{sigma_sq2}{Within-cluster variance for clusters in TREATMENT group}

\item{sigma_b_sq2}{Between-cluster variance for clusters in TREATMENT group.}

\item{alpha}{Significance level; default = 0.05.}

\item{quiet}{When set to FALSE, displays simulation progress and estimated completion time; default is FALSE.}

\item{all.sim.data}{Option to output list of all simulated datasets; default = FALSE.}

\item{seed}{Option to set seed. Default is NA.}

\item{poor.fit.override}{Option to override \code{stop()} if more than 25\% 
of fits fail to converge.
At least 2 of the following must be specified:}
}
\value{
A list with the following components:
\itemize{
  \item Character string indicating total number of simulations and simulation type
  \item Number of simulations
  \item Data frame with columns "Power" (Estimated statistical power), 
               "lower.95.ci" (Lower 95% confidence interval bound), 
               "upper.95.ci" (Upper 95% confidence interval bound)
  \item Analytic method used for power estimation
  \item Significance level
  \item Vector containing user-defined cluster sizes
  \item Vector containing user-defined number of clusters in each treatment group
  \item Data frame reporting ICC for Treatment/Non-Treatment groups
  \item Vector containing expected difference between groups based on user inputs
  \item Data frame with columns: 
                  "Estimate" (Estimate of treatment effect for a given simulation), 
                  "Std.err" (Standard error for treatment effect estimate), 
                  "Test.statistic" (z-value (for GLMM) or Wald statistic (for GEE)), 
                  "p.value", 
                  "sig.val" (Is p-value less than alpha?)
  \item List of data frames, each containing: 
                  "y" (Simulated response value), 
                  "trt" (Indicator for treatment group), 
                  "clust" (Indicator for cluster)
                  }
}
\description{
This function uses iterative simulations to determine 
approximate power for individually randomized group treatment trials with a 
normally-distributed outcome of interest. Users can modify a variety of 
parameters to suit the simulations to their desired experimental situation. 
This function returns the summary power values for each treatment arm.
}
\details{
Runs the power simulation.

Users must specify the desired number of simulations, number of subjects per 
cluster, number of clusters per treatment arm, expected absolute difference 
between treatments, two of the following: ICC, within-cluster variance, or 
between-cluster variance; significance level, analytic method, progress updates, 
and simulated data set output may also be specified.
}
\examples{
\dontrun{
irgtt.normal.sim <- cps.irgtt.normal(nsim = 100, nsubjects = 10, 
                       nclusters = 10, difference = .5,
                        sigma_sq = 1, sigma_sq2 = 0.9, 
                        sigma_b_sq2 = 0.1, alpha = 0.05,
                        quiet = FALSE, all.sim.data = FALSE)
}
}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu}), Alexander R. Bogdan, 
  and Ken Kleinman (\email{ken.kleinman@gmail.com})
}
