% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.ma.normal.internal.R
\name{cps.ma.normal.internal}
\alias{cps.ma.normal.internal}
\title{Power simulations for cluster-randomized trials: Simple Designs, Continuous Outcome.}
\usage{
cps.ma.normal.internal(nsim = 1000, str.nsubjects = NULL,
  means = NULL, sigma = NULL, sigma_b = NULL, alpha = 0.05,
  quiet = FALSE, method = "glmm", all.sim.data = FALSE,
  seed = NULL, poor.fit.override = FALSE)
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{str.nsubjects}{Number of subjects per treatment group; accepts a list with one entry per arm. 
Each entry is a vector containing the number of subjects per cluster (required).}

\item{means}{Expected absolute treatment effect for each arm; accepts a vector of length \code{narms} (required).}

\item{sigma}{Within-cluster variance; accepts a vector of length \code{narms} (required).}

\item{sigma_b}{Between-cluster variance; accepts a vector of length \code{narms} (required).}

\item{alpha}{Significance level; default = 0.05.}

\item{quiet}{When set to FALSE, displays simulation progress and estimated completion time; default is FALSE.}

\item{method}{Analytical method, either Generalized Linear Mixed Effects Model (GLMM) or 
Generalized Estimating Equation (GEE). Accepts c('glmm', 'gee') (required); default = 'glmm'.}

\item{all.sim.data}{Option to output list of all simulated datasets; default = FALSE.}

\item{seed}{Option to set.seed. Default is NULL.}

\item{poor.fit.override}{Option to override \code{stop()} if more than 25% of fits fail to converge}
}
\description{
This set of functions utilize iterative simulations to determine 
approximate power for cluster-randomized controlled trials. Users 
can modify a variety of parameters to suit the simulations to their
desired experimental situation.
}
\details{
Runs the power simulation.

Users must specify the desired number of simulations, number of subjects per 
cluster, number of clusters per treatment arm, group means, two of the following: ICC, within-cluster variance, or 
between-cluster variance; significance level, analytic method, progress updates, 
and simulated data set output may also be specified.
}
\examples{
\dontrun{

str.nsubjects.example <- list(c(20,20,20,25), c(15, 20, 20, 21), c(17, 20, 21))
means.example <- c(30, 21, 53)
sigma.example <- c(100, 110, 100)
sigma_b.example <- c(25, 25, 120)

normal.ma.rct <- cps.ma.normal.internal(nsim = 1000, str.nsubjects = str.nsubjects.example, 
                                       means = means.example, sigma = sigma.example, 
                                       sigma_b = sigma_b.example, alpha = 0.05, 
                                       quiet = FALSE, method = 'glmm', 
                                       all.sim.data = FALSE, seed = 123)
}

}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu}, Alexander R. Bogdan, and Ken Kleinman (\email{ken.kleinman@gmail.com})
}
