% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.ma.normal.internal.R
\name{cps.ma.normal.internal}
\alias{cps.ma.normal.internal}
\title{Runs the power simulations for cluster-randomized trials: Multi-Arm Design, Continuous Outcome.}
\usage{
cps.ma.normal.internal(nsim = 1000, str.nsubjects = NULL,
  means = NULL, sigma_sq = NULL, sigma_b_sq = NULL, alpha = 0.05,
  quiet = FALSE, method = "glmm", all.sim.data = FALSE,
  seed = NULL, cores = NULL, poor.fit.override = FALSE)
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{str.nsubjects}{Number of subjects per treatment group; accepts a list with one entry per arm. 
Each entry is a vector containing the number of subjects per cluster (required).}

\item{means}{Expected absolute treatment effect for each arm; accepts a vector of length \code{narms} (required).}

\item{sigma_sq}{Within-cluster variance; accepts a vector of length \code{narms} (required).}

\item{sigma_b_sq}{Between-cluster variance; accepts a vector of length \code{narms} (required).}

\item{alpha}{Significance level; default = 0.05.}

\item{quiet}{When set to FALSE, displays simulation progress and estimated completion time; default is FALSE.}

\item{method}{Analytical method, either Generalized Linear Mixed Effects Model (GLMM) or 
Generalized Estimating Equation (GEE); accepts c('glmm', 'gee') (required); default = 'glmm'.}

\item{all.sim.data}{Option to output list of all simulated datasets; default = FALSE.}

\item{seed}{Option to set.seed, default is NULL.}

\item{cores}{a string or numeric value indicating the number of cores to be used for parallel computing. 
When this option is set to NULL, no parallel computing is used.}

\item{poor.fit.override}{Option to override \code{stop()} if more than 25% of fits fail to converge}
}
\value{
A list with the following components:
\itemize{
  \item List of \code{length(nsim)} containing gee- or glmm-fitted the model summaries.
  \item Compares fitted model to a model for H0 using ML (anova).
  \item List of data frames, each containing: 
                  "y" (Simulated response value), 
                  "trt" (Indicator for treatment group), 
                  "clust" (Indicator for cluster)
  \item A vector of length \code{nsim} consisting of 1 and 0; 
}
}
\description{
This set of functions utilize iterative simulations to determine 
approximate power for cluster-randomized controlled trials. Users 
can modify a variety of parameters to suit the simulations to their
desired experimental situation. Users must specify the desired number of simulations, number of subjects per 
cluster, number of clusters per treatment arm, group means, two of the following: ICC, within-cluster variance, or 
between-cluster variance; significance level, analytic method, progress updates, 
and simulated data set output may also be specified.
}
\examples{
\dontrun{
str.nsubjects.example <- list(c(20,20,20,25), c(15, 20, 20, 21), c(17, 20, 21))
means.example <- c(30, 21, 53)
sigma_sq.example <- c(100, 110, 100)
sigma_b_sq.example <- c(25, 25, 120)
normal.ma.rct <- cps.ma.normal.internal(nsim = 1000, str.nsubjects = str.nsubjects.example, 
                                       means = means.example, sigma_sq = sigma_sq.example, 
                                       sigma_b_sq = sigma_b_sq.example, alpha = 0.05, 
                                       quiet = FALSE, method = 'glmm', 
                                       all.sim.data = FALSE, seed = 123)
                                       }
}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu}, Alexander R. Bogdan, and Ken Kleinman (\email{ken.kleinman@gmail.com})
}
