% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.normal.R
\name{cps.normal}
\alias{cps.normal}
\title{Power simulations for cluster-randomized trials: Simple Designs.}
\usage{
cps.normal(nsim = NULL, m = NULL, n = NULL, difference = NULL,
  ICC = NULL, sigma = NULL, sigma_b = NULL, ICC2 = NULL,
  sigma2 = NULL, sigma_b2 = NULL, alpha = 0.05, method = "glmm",
  quiet = FALSE)
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{m}{Number of subjects per cluster; accepts integer (required). Currently in development to
accept vectors of treatment-specific cluster numbers and distributional forms.}

\item{n}{Number of clusters per treatment group; accepts single integer (required) or 
vector of length 2 for unequal number of clusters per treatment group}

\item{difference}{Expected absolute treatment effect; accepts numeric (required).
At least 2 of the following must be specified:}

\item{ICC}{Intra-cluster correlation coefficient; accepts a value between 0 - 1}

\item{sigma_b}{Between-cluster variance; accepts numeric
If clusters differ between treatment groups, at least 2 of the following 
must be specified:}

\item{ICC2}{Intra-cluster correlation coefficient for clusters in TREATMENT group}

\item{sigma_b2}{Between-cluster variance for clusters in TREATMENT group}

\item{alpha}{Significance level. Default = 0.05.}

\item{method}{Analytical method, either Generalized Linear Mixed Effects Model (GLMM) or Generalized Estimating Equation (GEE). Accepts c('glmm', 'gee') (required); default = 'glmm'.}

\item{quiet}{When set to FALSE, displays simulation progress and estimated completion time. Default is FALSE.}

\item{sigma_w}{Within-cluster variance; accepts numeric}

\item{sigma_w2}{Within-cluster variance for clusters in TREATMENT group}
}
\description{
This set of functions utilize iterative simulations to determine 
approximate power for cluster-randomized controlled trials. Users 
can modify a variety of parameters to suit the simulations to their
desired experimental situation.
}
\details{
Runs the power simulation.

Users must specify the desired number of simulations, number of subjects per 
cluster, number of clusters per treatment arm, expected absolute difference 
between treatments, two of the following: ICC, within-cluster variance, or 
between-cluster variance; significance level, analytic method, and whether 
or not progress updates should be displayed while the function is running.
}
\examples{
\dontrun{
my.normal.sim = cps.normal(nsim = 100, m = 50, n = 30, difference = 30, ICC = 0.2, sigma_w = 100,
                    alpha = 0.05, method = 'glmm', quiet = FALSE)
}

}
\author{
Alexander R. Bogdan
}
