% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.normal.R
\name{cps.normal}
\alias{cps.normal}
\title{Power simulations for cluster-randomized trials: Parallel Designs, Normal Outcome}
\usage{
cps.normal(
  nsim = NULL,
  nclusters = NULL,
  nsubjects = NULL,
  mu = 0,
  mu2 = NULL,
  ICC = NULL,
  sigma_sq = NULL,
  sigma_b_sq = NULL,
  ICC2 = NULL,
  sigma_sq2 = NULL,
  sigma_b_sq2 = NULL,
  alpha = 0.05,
  method = "glmm",
  quiet = FALSE,
  all.sim.data = FALSE,
  seed = NA,
  poor.fit.override = FALSE,
  irgtt = FALSE,
  nofit = FALSE
)
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{nclusters}{Number of clusters per group; accepts single integer (implying equal numbers of clusters in the two groups)
or vector of length 2 (unequal number of clusters per arm) (required)}

\item{nsubjects}{Number of subjects per cluster; accepts either a scalar (implying equal cluster sizes for the two groups), 
a vector of length two (equal cluster sizes within groups), or a vector of length \code{sum(nclusters)} 
(unequal cluster sizes within groups) (required).}

\item{mu}{Mean in the first arm; accepts numeric, default 0 (required).}

\item{mu2}{Mean in the second arm; accepts numeric (required).

At least 2 of the following must be specified:}

\item{ICC}{Intra-cluster correlation coefficient; accepts a value between 0 - 1}

\item{sigma_sq}{Within-cluster variance; accepts numeric}

\item{sigma_b_sq}{Between-cluster variance; accepts numeric


The defaults for the following are all NA, implying equal variance parameters 
for the two groups. If one of the following is given, variance parameters differ 
between treatment groups, and at least 2 of the following 
must be specified:}

\item{ICC2}{Intra-cluster correlation coefficient for clusters in the second group}

\item{sigma_sq2}{Within-cluster variance for clusters in the second group}

\item{sigma_b_sq2}{Between-cluster variance for clusters in the second group

Optional parameters:}

\item{alpha}{Significance level; default = 0.05.}

\item{method}{Analytical method, either Generalized Linear Mixed Effects Model (GLMM) or 
Generalized Estimating Equation (GEE). Accepts c('glmm', 'gee') (required); default = 'glmm'.}

\item{quiet}{When set to FALSE, displays simulation progress and estimated completion time; default is FALSE.}

\item{all.sim.data}{Option to output list of all simulated datasets; default = FALSE.}

\item{seed}{Option to set the seed. Default is NA.}

\item{poor.fit.override}{Option to override \code{stop()} if more than 25\% 
of fits fail to converge.}

\item{irgtt}{Logical. Is the experimental design an individually randomized 
group treatment trial? For details, see ?cps.irgtt.normal.}

\item{nofit}{Option to return only the simulated data, no analysis. Defaults to FALSE.}
}
\value{
If \code{nofit = FALSE} then a list with the following components:
\itemize{
  \item Character string indicating total number of simulations and simulation type
  \item Number of simulations
  \item Data frame with columns "Power" (Estimated statistical power), 
               "lower.95.ci" (Lower 95% confidence interval bound), 
               "upper.95.ci" (Upper 95% confidence interval bound).
               Note that non-convergent models are returned for review, 
               but not included in this calculation.
  \item Analytic method used for power estimation
  \item Significance level
  \item Vector containing user-defined cluster sizes
  \item Vector containing user-defined number of clusters in each arm
  \item Data frame reporting ICC, variance parameters, and means for each arm
  \item Vector containing expected group means based on user inputs
  \item Data frame with columns: 
                  "Estimate" (Estimate of treatment effect for a given simulation), 
                  "Std.err" (Standard error for treatment effect estimate), 
                  "Test.statistic" (z-value (for GLMM) or Wald statistic (for GEE)), 
                  "p.value", 
                  "sig.val" (Is p-value less than alpha?)
  \item List of data frames, each containing: 
                  "y" (Simulated response value), 
                  "trt" (Indicator for arm), 
                  "clust" (Indicator for cluster)
                  }

If \code{nofit = TRUE} then a dataframe with \code{nsim} + 2 columns and a row for 
each simulated subject.
}
\description{
\loadmathjax

This function uses Monte Carlo methods (simulations) to estimate 
power for cluster-randomized trials. Users 
can modify a variety of parameters to suit the simulations to their
desired experimental situation.

Users must specify the desired number of simulations, number of subjects per 
cluster, number of clusters per arm, expected means of the arms, and two of 
the following: ICC, within-cluster variance, or between-cluster variance.  
Defaults are provided for significance level, analytic method, progress 
updates, and whether the simulated data sets are retained.

Users have the option of specifying different variance parameters for each
group, different numbers of clusters for each treatmetnt group, and different numbers
of units within each cluster. 

Non-convergent models are not included in the calculation of exact confidence 
intervals.
}
\section{Testing details}{
   
This function has been verified, where possible, against reference values from the NIH's GRT 
Sample Size Calculator, PASS11, \code{CRTsize::n4means}, and 
\code{clusterPower::cpa.normal}.
}

\section{Notes}{


The data generating model here is:
\mjsdeqn{y_{ij} = \mu + b_i + e_{ij} }
for the first group or arm, where \mjseqn{b_i \sim N(0,\sigma_b^2)} 
and \mjseqn{e_{ij} \sim N(0,\sigma^2)}, while for the second group, 
 
\mjsdeqn{y_{ij} = \mu_2 + b_i + e_{ij} }
where \mjseqn{b_i \sim N(0,\sigma_{b_2}^2)} 
and \mjseqn{e_{ij} \sim N(0,\sigma_2^2)}; if none of 
\mjseqn{\sigma_2^2, \sigma_{b_2}^2} or \code{ICC2} are used, then the second group uses
\mjseqn{b_i \sim N(0,\sigma_b^2)} 
and \mjseqn{e_{ij} \sim N(0,\sigma^2)}.

All random terms are generated indepent of one another.


For calls without \mjseqn{\sigma_2^2, \sigma_{b_2}^2} or \code{ICC2}, and using
\code{method="glmm"} the fitted model is:
\mjsdeqn{y_{ij}|b_i = \mu + \beta_1 x_{ij}  + b_i + e_{ij}}

with \mjseqn{\beta_1 = \mu_2 - \mu},
treatment group indicator \mjseqn{x_{ij} = 0} for the first group, with
\mjseqn{b_i \sim N(0, \sigma_b^2)} and \mjseqn{e_{ij} \sim N(0,\sigma^2)},
the random effects distribution and and residual distribution common for both
treatment groups.

Otherwise, for \code{method="glmm"} the fitted model is:
\mjsdeqn{y_{ij}|b_i = \mu + \beta_1 x_{ij} 
   + b_i I(x_{ij}=0) + e_{ij} I(x_{ij}=0) 
   + g_i I(x_{ij}=1) + f_{ij} I(x_{ij}=1)
   }

with \mjseqn{\beta_1}, \mjseqn{x_{ij}, b_i}, and \mjseqn{e_{ij}} as above, with
\mjseqn{g_i \sim N(0, \sigma_{b_2}^2)} and \mjseqn{f \sim N(0,\sigma_2^2)}, the 
random effects and residual distribution in the second group.
}

\examples{

# Estimate power for an trial with 10 clusters in each arm and 50 subjects in each 
# cluster, with an ICC of .3, sigma squared of 20 (implying sigma_b squared of 8.57143) 
# in each group, with arm means of 1 and 4.75 in the two groups using 100 simulated 
# data sets.  
   
\dontrun{

normal.sim = cps.normal(nsim = 100, nsubjects = 25, nclusters = 10, mu = 0, 
  mu2 = 0.5, ICC = 0.05, sigma_sq = 2)
}

# Estimate power for a trial with 5 clusters in one arm, those clusters having 100 subjects 
# each, 25 clusters in the other arm, those clusters having 50 subjects each, the fist arm
# having a sigma squared of 20 and sigma_b squared of 8.57143, and the second a sigma squared
# of 9 and a sigma_b squared of 1, with estimated arm means of 1 and 4.75 in the first and 
# second groups, respectively, using 100 simulated data sets.

\dontrun{

normal.sim2 = cps.normal(nsim = 100, nclusters = c(5,25), nsubjects = c(100,50), mu = 1, 
  mu2 = 4.75, sigma_sq = 20,sigma_b_sq = 8.8571429, sigma_sq2 = 9, sigma_b_sq2 = 1)
}


}
\author{
Alexander R. Bogdan, Alexandria C. Sakrejda 
(\email{acbro0@umass.edu}), and Ken Kleinman 
(\email{ken.kleinman@gmail.com})
}
