% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.normal.R
\name{cps.normal}
\alias{cps.normal}
\title{Power simulations for cluster-randomized trials: Simple Designs, Continuous Outcome.}
\usage{
cps.normal(nsim = NULL, nsubjects = NULL, nclusters = NULL,
  difference = NULL, ICC = NULL, sigma = NULL, sigma_b = NULL,
  ICC2 = NULL, sigma2 = NULL, sigma_b2 = NULL, alpha = 0.05,
  method = "glmm", quiet = FALSE, all.sim.data = FALSE, seed = NA,
  irgtt = FALSE)
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{nsubjects}{Number of subjects per cluster; accepts either a scalar (equal cluster sizes, both groups), 
a vector of length two (equal cluster sizes within groups), or a vector of length \code{sum(nclusters)} 
(unequal cluster sizes within groups) (required).}

\item{nclusters}{Number of clusters per group; accepts single integer or vector of length 2 for unequal number 
of clusters per treatment group (required)}

\item{difference}{Expected absolute treatment effect; accepts numeric (required).
At least 2 of the following must be specified:}

\item{ICC}{Intra-cluster correlation coefficient; accepts a value between 0 - 1}

\item{sigma}{Within-cluster variance; accepts numeric}

\item{sigma_b}{Between-cluster variance; accepts numeric

If clusters differ between treatment groups, at least 2 of the following 
must be specified:}

\item{ICC2}{Intra-cluster correlation coefficient for clusters in TREATMENT group}

\item{sigma2}{Within-cluster variance for clusters in TREATMENT group}

\item{sigma_b2}{Between-cluster variance for clusters in TREATMENT group}

\item{alpha}{Significance level; default = 0.05.}

\item{method}{Analytical method, either Generalized Linear Mixed Effects Model (GLMM) or 
Generalized Estimating Equation (GEE). Accepts c('glmm', 'gee') (required); default = 'glmm'.}

\item{quiet}{When set to FALSE, displays simulation progress and estimated completion time; default is FALSE.}

\item{all.sim.data}{Option to output list of all simulated datasets; default = FALSE.}

\item{seed}{Option to set the seed. Default is NA.}
}
\value{
A list with the following components:
\itemize{
  \item Character string indicating total number of simulations and simulation type
  \item Number of simulations
  \item Data frame with columns "Power" (Estimated statistical power), 
               "lower.95.ci" (Lower 95% confidence interval bound), 
               "upper.95.ci" (Upper 95% confidence interval bound)
  \item Analytic method used for power estimation
  \item Significance level
  \item Vector containing user-defined cluster sizes
  \item Vector containing user-defined number of clusters in each treatment group
  \item Data frame reporting ICC for Treatment/Non-Treatment groups
  \item Vector containing expected difference between groups based on user inputs
  \item Data frame with columns: 
                  "Estimate" (Estimate of treatment effect for a given simulation), 
                  "Std.err" (Standard error for treatment effect estimate), 
                  "Test.statistic" (z-value (for GLMM) or Wald statistic (for GEE)), 
                  "p.value", 
                  "sig.val" (Is p-value less than alpha?)
  \item List of data frames, each containing: 
                  "y" (Simulated response value), 
                  "trt" (Indicator for treatment group), 
                  "clust" (Indicator for cluster)
                  }
}
\description{
This set of functions utilize iterative simulations to determine 
approximate power for cluster-randomized controlled trials. Users 
can modify a variety of parameters to suit the simulations to their
desired experimental situation.
}
\details{
Runs the power simulation.

Users must specify the desired number of simulations, number of subjects per 
cluster, number of clusters per treatment arm, expected absolute difference 
between treatments, two of the following: ICC, within-cluster variance, or 
between-cluster variance; significance level, analytic method, progress updates, 
and simulated data set output may also be specified.
}
\examples{
\dontrun{
normal.sim = cps.normal(nsim = 100, nsubjects = 50, nclusters = 9, difference = 10,
                        ICC = 0.3, sigma = 100, alpha = 0.05, method = 'glmm', 
                        quiet = FALSE, all.sim.data = FALSE)
}
}
\author{
Alexander R. Bogdan, Alexandria C. Sakrejda 
(\email{acbro0@umass.edu}), and Ken Kleinman 
(\email{ken.kleinman@gmail.com})
}
