% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.sw.binary.R
\name{cps.sw.binary}
\alias{cps.sw.binary}
\title{Power simulations for cluster-randomized trials: Stepped Wedge Design, Binary Outcome}
\usage{
cps.sw.binary(nsim = NULL, nsubjects = NULL, nclusters = NULL,
  p.ntrt = NULL, p.trt = NULL, steps = NULL, sigma_b = NULL,
  alpha = 0.05, method = "glmm", quiet = FALSE,
  all.sim.data = FALSE)
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{nsubjects}{Number of subjects per cluster; accepts either a scalar (equal cluster sizes) 
or a vector of length \code{nclusters} (user-defined size for each cluster) (required).}

\item{nclusters}{Number of clusters; accepts non-negative integer scalar (required).}

\item{p.ntrt}{Expected probability of outcome in non-treatment group. Accepts scalar between 0 - 1 (required).}

\item{p.trt}{Expected probability of outcome in treatment group. Accepts scalar between 0 - 1 (required).}

\item{steps}{Number of crossover steps; a baseline step (all clusters in non-treatment group) is assumed. 
Accepts positive scalar (indicating the total number of steps; clusters per step is obtained by 
\code{nclusters / steps}) or a vector of non-negative integers corresponding either to the number 
of clusters to be crossed over at each time point (e.g c(2,4,4,2); nclusters = 10) or the cumulative 
number of clusters crossed over by a given time point (e.g. c(2,4,8,10); nclusters = 10) (required).}

\item{sigma_b}{Between-cluster variance; accepts non-negative numeric scalar (indicating equal 
between-cluster variances for both treatment groups) or a vector of length 2 specifying treatment-specific 
between-cluster variances (required).}

\item{alpha}{Significance level. Default = 0.05.}

\item{method}{Analytical method, either Generalized Linear Mixed Effects Model (GLMM) or 
Generalized Estimating Equation (GEE). Accepts c('glmm', 'gee') (required); default = 'glmm'.}

\item{quiet}{When set to FALSE, displays simulation progress and estimated completion time; default is FALSE.}

\item{all.sim.data}{Option to output list of all simulated datasets; default = FALSE.}
}
\description{
This set of functions utilize iterative simulations to determine 
approximate power for stepped wedge cluster-randomized controlled trials. Users 
can modify a variety of parameters to suit the simulations to their
desired experimental situation.
}
\details{
Runs power simulations for stepped wedge cluster-randomized controlled trials with a binary outcome.

Users must specify the desired number of simulations, number of subjects per 
cluster, number of clusters per treatment arm, expected absolute difference 
between treatment arms, within-cluster variance, between-cluster variance, 
significance level, analytic method, progress updates, and simulated data 
set output may also be specified.
}
\examples{
\dontrun{
binary.sw.rct = cps.sw.binary(nsim = 100, nsubjects = 50, nclusters = 30, 
                              p.ntrt = 0.1, p.trt = 0.2, steps = 5, 
                              sigma_b = 30, alpha = 0.05, method = 'glmm', 
                              quiet = FALSE, all.sim.data = FALSE)
}

}
\author{
Alexander R. Bogdan
}
