% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.sw.binary.R
\name{cps.sw.binary}
\alias{cps.sw.binary}
\title{Power simulations for cluster-randomized trials: Stepped Wedge Design, Binary Outcome}
\usage{
cps.sw.binary(nsim = NULL, nsubjects = NULL, nclusters = NULL,
  p.ntrt = NULL, p.trt = NULL, steps = NULL, sigma_b_sq = NULL,
  alpha = 0.05, method = "glmm", quiet = FALSE,
  all.sim.data = FALSE, opt = "L-BFGS-B")
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{nsubjects}{Number of subjects per cluster; accepts either a scalar (equal cluster sizes) 
or a vector of length \code{nclusters} (user-defined size for each cluster) (required).}

\item{nclusters}{Number of clusters; accepts non-negative integer scalar (required).}

\item{p.ntrt}{Expected probability of outcome in non-treatment group. Accepts scalar between 0 - 1 (required).}

\item{p.trt}{Expected probability of outcome in treatment group. Accepts scalar between 0 - 1 (required).}

\item{steps}{Number of crossover steps; a baseline step (all clusters in non-treatment group) is assumed. 
Accepts positive scalar (indicating the total number of steps; clusters per step is obtained by 
\code{nclusters / steps}) or a vector of non-negative integers corresponding either to the number 
of clusters to be crossed over at each time point (e.g c(2,4,4,2); nclusters = 10) or the cumulative 
number of clusters crossed over by a given time point (e.g. c(2,4,8,10); nclusters = 10) (required).}

\item{sigma_b_sq}{Between-cluster variance; accepts non-negative numeric scalar (indicating equal 
between-cluster variances for both treatment groups) or a vector of length 2 specifying treatment-specific 
between-cluster variances (required).}

\item{alpha}{Significance level. Default = 0.05.}

\item{method}{Analytical method, either Generalized Linear Mixed Effects Model (GLMM) or 
Generalized Estimating Equation (GEE). Accepts c('glmm', 'gee') (required); default = 'glmm'.}

\item{quiet}{When set to FALSE, displays simulation progress and estimated completion time; default is FALSE.}

\item{all.sim.data}{Option to output list of all simulated datasets; default = FALSE.}

\item{opt}{Option to fit with a different optimizer (using the package \code{optimx}). Default is 'L-BFGS-B'.}
}
\value{
A list with the following components
\itemize{
  \item Character string indicating total number of simulations and simulation type
  \item Number of simulations
  \item Data frame with columns "Power" (Estimated statistical power), 
               "lower.95.ci" (Lower 95% confidence interval bound), 
               "upper.95.ci" (Upper 95% confidence interval bound)
  \item Analytic method used for power estimation
  \item Significance level
  \item Vector containing user-defined cluster sizes
  \item Vector containing user-defined number of clusters
  \item Data frame reporting ICC, within & between cluster variances for Treatment/Non-Treatment
   groups at each time point
  \item Vector containing expected difference between groups based on user inputs
  \item Data frame containing mean response values for each treatment group at each time point
  \item Matrix showing cluster crossover at each time point
  \item Data frame with columns: 
                  "Estimate" (Estimate of treatment effect for a given simulation), 
                  "Std.err" (Standard error for treatment effect estimate), 
                  "Test.statistic" (z-value (for GLMM) or Wald statistic (for GEE)), 
                  "p.value", 
                  "sig.val" (Is p-value less than alpha?)
  \item List of data frames, each containing: 
                  "y" (Simulated response value), 
                  "trt" (Indicator for treatment group),
                  "time.point" (Indicator for step; "t1" = time point 0) 
                  "clust" (Indicator for cluster), 
                  "period" (Indicator for at which step a cluster crosses over)
}
}
\description{
This set of functions utilize iterative simulations to determine 
approximate power for stepped wedge cluster-randomized controlled trials. Users 
can modify a variety of parameters to suit the simulations to their
desired experimental situation.
}
\details{
Runs power simulations for stepped wedge cluster-randomized controlled trials 
with a binary outcome. The stepped wedge trial design is a type of cross-over
design in which clusters change treatments in waves. Initially all the 
clusters recieve the same standard treatment, and at the end of the trial all
of the clusters will be recieving the treatment of interest. More than one 
cluster can change treatments in a wave, but the order in which clusters 
change treatments is randomly determined. The outcome of interest is assessed 
in each cluster during each wave.

Users must specify the desired number of simulations, number of subjects per 
cluster, number of clusters per treatment arm, expected absolute difference 
between treatment arms, within-cluster variance, between-cluster variance, 
significance level, analytic method, progress updates, and simulated data 
set output may also be specified.
}
\examples{
\dontrun{
binary.sw.rct = cps.sw.binary(nsim = 100, nsubjects = 50, nclusters = 30, 
                              p.ntrt = 0.1, p.trt = 0.2, steps = 5, 
                              sigma_b_sq = 30, alpha = 0.05, method = 'glmm', 
                              quiet = FALSE, all.sim.data = FALSE)
}

}
\author{
Alexander R. Bogdan

Alexandria C. Sakrejda (\email{acbro0@umass.edu})

Ken Kleinman (\email{ken.kleinman@gmail.com})
}
