% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.sw.count.R
\name{cps.sw.count}
\alias{cps.sw.count}
\title{Power simulations for cluster-randomized trials: Stepped Wedge Design, Count Outcome}
\usage{
cps.sw.count(nsim = NULL, nsubjects = NULL, nclusters = NULL,
  c.ntrt = NULL, c.trt = NULL, steps = NULL, sigma_b = NULL,
  alpha = 0.05, family = "poisson", analysis = "poisson",
  method = "glmm", quiet = FALSE, all.sim.data = FALSE)
}
\arguments{
\item{nsim}{Number of datasets to simulate; accepts integer (required).}

\item{nsubjects}{Number of subjects per cluster; accepts either a scalar (equal cluster sizes) 
or a vector of length \code{nclusters} (user-defined size for each cluster) (required).}

\item{nclusters}{Number of clusters; accepts non-negative integer scalar (required).}

\item{c.ntrt}{Expected outcome count in non-treatment group. Accepts scalar between 0 - 1 (required).}

\item{c.trt}{Expected outcome count in treatment group. Accepts scalar between 0 - 1 (required).}

\item{steps}{Number of crossover steps; a baseline step (all clusters in non-treatment group) is assumed. 
Accepts positive scalar (indicating the total number of steps; clusters per step is obtained by 
\code{nclusters / steps}) or a vector of non-negative integers corresponding either to the number 
of clusters to be crossed over at each time point (e.g c(2,4,4,2); nclusters = 10) or the cumulative 
number of clusters crossed over by a given time point (e.g. c(2,4,8,10); nclusters = 10) (required).}

\item{sigma_b}{Between-cluster variance; accepts non-negative numeric scalar (indicating equal 
between-cluster variances for both treatment groups) or a vector of length 2 specifying treatment-specific 
between-cluster variances (required).}

\item{alpha}{Significance level. Default = 0.05.}

\item{family}{Distribution from which responses are simulated. Accepts Poisson ('poisson') or negative binomial ('neg.binom') (required); default = 'poisson'}

\item{analysis}{Family used for regression; currently only applicable for GLMM. Accepts c('poisson', 'neg.binom') (required); default = 'poisson'}

\item{method}{Analytical method, either Generalized Linear Mixed Effects Model (GLMM) or 
Generalized Estimating Equation (GEE). Accepts c('glmm', 'gee') (required); default = 'glmm'.}

\item{quiet}{When set to FALSE, displays simulation progress and estimated completion time; default is FALSE.}

\item{all.sim.data}{Option to output list of all simulated datasets; default = FALSE.}
}
\description{
This set of functions utilize iterative simulations to determine 
approximate power for stepped wedge cluster-randomized controlled trials. Users 
can modify a variety of parameters to suit the simulations to their
desired experimental situation.
}
\details{
Runs power simulations for stepped wedge cluster-randomized controlled trials with a count outcome.

Users must specify the desired number of simulations, number of subjects per 
cluster, number of clusters per treatment arm, expected absolute difference 
between treatment arms, between-cluster variance; significance level, 
distributional family for data generation, analytic method, progress updates, 
and simulated data set output may also be specified.
}
\examples{
\dontrun{
count.sw.rct = cps.sw.count(nsim = 100, nsubjects = 50, nclusters = 30, 
                              c.ntrt = 3, c.trt = 5, steps = 5, sigma_b = 20, 
                              alpha = 0.05, family = 'poisson', analysis = 'poisson', 
                              method = 'glmm', quiet = FALSE, all.sim.data = FALSE)
}

}
\author{
Alexander R. Bogdan
}
