% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMissingVarianceParam.R
\name{createMissingVarianceParam}
\alias{createMissingVarianceParam}
\title{Power simulations for cluster-randomized trials: calculating ICC, 
sigma_sq, and sigma_b_sq inputs}
\usage{
createMissingVarianceParam(ICC = NULL, sigma_sq = NULL,
  sigma_b_sq = NULL)
}
\arguments{
\item{ICC}{Intra-cluster correlation coefficient; accepts a vector of length \code{narms} with values between 0 - 1.}

\item{sigma_sq}{Within-cluster variance; accepts a vector of length \code{narms}.}

\item{sigma_b_sq}{Between-cluster variance; accepts a vector of length \code{narms}.}
}
\value{
A vector of length \code{narms} containing the argument that the user did not specify.
}
\description{
Usually called from within a function, createMissingVarianceParam takes 2 of 3 
arguments (ICC, sigma_sq, or sigma_b_sq) and returns the missing argument. It gives 
an error if one of the inputs is missing or not specified properly. Note that 
it is required that the user specify at least 2 of the arguments.
}
\examples{
\dontrun{
ICC <- createMissingVarianceParam(sigma_sq = c(1, 1, 0.9), sigma_b_sq = c(0.1, 0.15, 0.1))
}

}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu}, Alexander R. Bogdan, and Ken Kleinman (\email{ken.kleinman@gmail.com})
}
