% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationEngines.R
\name{mixed.eff.params}
\alias{mixed.eff.params}
\title{Calculation of variance in Poisson mixed model setting.}
\usage{
mixed.eff.params(pi0, btw.clust.var, Tk)
}
\arguments{
\item{pi0}{the baseline cluster-level mean on the scale of the link function}

\item{btw.clust.var}{the between-cluster-variance}

\item{Tk}{the at-risk time for each cluster}
}
\value{
A numeric vector with the following three named elements, in order:
  ["expectation"] the overall mean of cluster-level outcomes, ["variance"] 
  the overall variance of cluster-level outcomes, ["hayes.k"] the estimated
  coefficient of variation.
}
\description{
This function is designed to calculate the overall variance for cluster-level 
  outcomes in a mixedeffect Poisson model. Conditional expectation calculations
  are implemented.
  
\code{mixed.eff.params()} is used by the \code{hayes.power.poisson()} function to 
  compute the effective coefficient of variation, or k, for a particular
  study design.
}
\examples{
mixed.eff.params(pi0=log(1), btw.clust.var=.5, Tk=100)
}
