% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_H0_rejection.R
\name{prop_H0_rejection}
\alias{prop_H0_rejection}
\title{Power simulations for cluster-randomized trials: proportion of F-test rejections}
\usage{
prop_H0_rejection(
  alpha = alpha,
  nsim = nsim,
  LRT.holder.abbrev = LRT.holder.abbrev
)
}
\arguments{
\item{alpha}{A numeric; the user-selected alpha cutoff.}

\item{nsim}{A scalar; the number of simulations chosen by the user.}

\item{LRT.holder.abbrev}{A scalar; the number of test rejections.}
}
\value{
A dataframe
\describe{
  \item{Ftest}{Exact confidence intervals produced using \code{binom.test()}}
}
}
\description{
This function can be called by some of the clusterPower functions to 
calculate the proportion of F-test rejections and produce exact confidence intervals.
}
\author{
Alexandria C. Sakrejda (\email{acbro0@umass.edu} and Ken Kleinman (\email{ken.kleinman@gmail.com})
}
