% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run a Shiny app for power analysis functions}
\usage{
runExample()
}
\description{
This function runs a Shiny app that employs the \code{crtpwr.2mean},
\code{crtpwr.2prop}, and \code{crtpwr.2rate}. The app can do calculations
for several sets of parameters and produces graphs.
}
\section{Authors}{

Jonathan Moyer (\email{jon.moyer@gmail.com})
}

